### 2023-10-24 #5965 @kyleniemeyer
Hi @openjournals/dev, it seems the wordcount function can find the paper, but something with generating the PDF is failing. Any ideas?
