### 2023-10-25 #5965 @xuanxu
For the PDF to compile the source file should be named `paper.md`
