+++ paper/FINE-literature.bib
@@ -76,7 +76,8 @@ volume={3},
 number={3},
 pages={219--260},
 year={2011},
-publisher={Springer}
+publisher={Springer},
+DOI = {10.1007/s12532-011-0026-8}
 }
 
 @article{Hoffmann_2020,
@@ -95,8 +96,10 @@ publisher={Springer}
 @article{Hoffmann_2021,
 title={Typical periods or typical time steps? A multi-model analysis to determine the optimal temporal aggregation for energy system models},
 author={Hoffmann, Maximilian and Priesmann, Jan and Nolting, Lars and Praktiknjo, Aaron and Kotzur, Leander and Stolten, Detlef},
-journal = {Preprints},
-DOI = {arXiv:2103.16657},
+journal = {Applied Energy},
+volume = {304},
+issn = {0306-2619},
+DOI = {10.1016/j.apenergy.2021.117825},
 year = {2021},
 type = {Journal Article}
 }
