+++ requirements_dev.yml
@@ -16,7 +16,7 @@ dependencies:
   - conda-forge::xlrd
   - conda-forge::pyomo
   - conda-forge::numpy
-  - conda-forge::pandas>=2
+  - conda-forge::pandas
   - conda-forge::networkx
   - conda-forge::scipy
   - conda-forge::scikit-learn
