### 2023-11-16 088610e
update subclass conversionPartLoad
