+++ FINE/subclasses/conversionPartLoad.py
@@ -1,1068 +1,1071 @@
-# from FINE.conversion import Conversion, ConversionModel
-# from FINE.utils import checkDataFrameConversionFactor, checkCallableConversionFactor
-# from FINE import utils
-# import pyomo.environ as pyomo
-# import pandas as pd
-# import numpy as np
-# import warnings
-
-
-# def pieceWiseLinearization(functionOrRaw, xLowerBound, xUpperBound, nSegments):
-#     """
-#     Determine xSegments, ySegments.
-#     If nSegments is not specified by the user it is either set (e.g. nSegments=5) or nSegements is determined by
-#     a bayesian optimization algorithm.
-#     """
-#     try:
-#         from GPyOpt.methods import BayesianOptimization
-#         import pwlf
-#     except ImportError as e:
-#         warnings.warn(
-#             """
-#             In order to use the `conversionPartLoadClass` you need to install `GPyOpt` and `pwlf`.
-#             GPyOpt reached end of maintenance and does not work with current versions of e.g. pandas.
-#             Make sure to downgrade necessary packages to make GPyOpt work for your Python installation.
-#             """,
-#             DeprecationWarning,
-#         )
-#         raise e
-
-#     if callable(functionOrRaw):
-#         nPointsForInputData = 1000
-#         x = np.linspace(xLowerBound, xUpperBound, nPointsForInputData)
-#         y = np.array([functionOrRaw(x_i) for x_i in x])
-#     else:
-#         x = np.array(functionOrRaw.iloc[:, 0])
-#         y = np.array(functionOrRaw.iloc[:, 1])
-#         if not 0.0 in x:
-#             xMinDefined = np.amin(x)
-#             xMaxDefined = np.amax(x)
-#             lenIntervalDefined = xMaxDefined - xMinDefined
-#             lenIntervalUndefined = xMinDefined
-#             nPointsUndefined = lenIntervalUndefined * (x.size / lenIntervalDefined)
-#             xMinIndex = np.argmin(x)
-#             for i in range(int(nPointsUndefined)):
-#                 x = np.append(x, [i / int(nPointsUndefined + 1) * lenIntervalUndefined])
-#                 y = np.append(y, y[xMinIndex])
-#         if not 1.0 in x:
-#             xMinDefined = np.amin(x)
-#             xMaxDefined = np.amax(x)
-#             lenIntervalDefined = xMaxDefined - xMinDefined
-#             lenIntervalUndefined = 1.0 - xMaxDefined
-#             nPointsUndefined = lenIntervalUndefined * (x.size / lenIntervalDefined)
-#             xMaxIndex = np.argmax(x)
-#             for i in range(int(nPointsUndefined)):
-#                 x = np.append(
-#                     x,
-#                     [
-#                         xMaxDefined
-#                         + (i + 1) / int(nPointsUndefined) * lenIntervalUndefined
-#                     ],
-#                 )
-#                 y = np.append(y, y[xMaxIndex])
-
-#     myPwlf = pwlf.PiecewiseLinFit(x, y)
-
-#     if nSegments == None:
-#         nSegments = 5
-#     elif nSegments == "optimizeSegmentNumbers":
-#         bounds = [{"name": "var_1", "type": "discrete", "domain": np.arange(2, 8)}]
-
-#         # Define objective function to get optimal number of line segments
-#         def my_obj(_x):
-#             # The penalty parameter l is set arbitrarily.
-#             # It depends upon the noise in your data and the value of your sum of square of residuals
-#             l = y.mean() * 0.001
-
-#             f = np.zeros(_x.shape[0])
-
-#             for i, j in enumerate(_x):
-#                 myPwlf.fit(j[0])
[…]
