+++ requirements_dev.yml
@@ -31,6 +31,7 @@ dependencies:
   - conda-forge::pytest-cov
   - conda-forge::nbval
   - conda-forge::black
-
+  - pwlf
+  
   - pip:
       - -e .
