+++ .gitlab-ci.yml
@@ -2,22 +2,51 @@ image: mambaorg/micromamba
 
 stages:
   - test
-  - prebuild
   - build
 
 variables:
   DOCKER_HOST: tcp://docker:2375
   DOCKER_TLS_CERTDIR: ""
 
+# Templates
+
 .test_template:
   stage: test
   before_script:
     - micromamba install -n base -y --file=requirements_dev.yml
+  rules:
+    # Run only for pushes to master or develop and for merge requests to master
+    - if: '$CI_COMMIT_BRANCH == "master"'
+    - if: '$CI_COMMIT_BRANCH == "develop"'
+    - if: '$CI_MERGE_REQUEST_TARGET_BRANCH_NAME == "master"'
+    # Switch from branch pipeline to merge pipeline once a merge request has
+    # been created on the branch.
+    - if: $CI_COMMIT_BRANCH && $CI_OPEN_MERGE_REQUESTS && $CI_PIPELINE_SOURCE == "push"
+      when: never
+    - if: $CI_COMMIT_BRANCH && $CI_OPEN_MERGE_REQUESTS
+      when: never
 
-.test_formatting_template:
+
+.test_docker_template:
   stage: test
+  image: jugit-registry.fz-juelich.de/iek-3/shared-code/fine/fine-dev:latest 
   before_script:
-    - micromamba install -y -c conda-forge "black"  "python=3.10"
+    - micromamba update -n base -y --file=requirements_dev.yml
+  rules:
+    # Do not run for pushes to master or develop and for merge requests to master
+    - if: $CI_COMMIT_BRANCH == "master"
+      when: never
+    - if: $CI_COMMIT_BRANCH == "develop"
+      when: never
+    - if: $CI_MERGE_REQUEST_TARGET_BRANCH_NAME == "master"
+      when: never
+    # Switch from branch pipeline to merge pipeline once a merge request has
+    # been created on the branch.
+    - if: $CI_COMMIT_BRANCH && $CI_OPEN_MERGE_REQUESTS && $CI_PIPELINE_SOURCE == "push"
+      when: never
+    - if: $CI_COMMIT_BRANCH && $CI_OPEN_MERGE_REQUESTS
+      when: never
+    - when: on_success
 
 .build_template:
   stage: build
@@ -25,44 +54,18 @@ variables:
   services:
     - docker:dind@sha256:c8bb6fa5388b56304dd770c4bc0478de81ce18540173b1a589178c0d31bfce90
   when: on_success
-  dependencies:
-    - prebuild
+
+# Tests
 
 test-code:
-  #  Switch from branch pipeline to merge pipeline once a merge request has been
-  #  created on the branch.
-  rules:
-    - if: $CI_COMMIT_BRANCH && $CI_OPEN_MERGE_REQUESTS && $CI_PIPELINE_SOURCE == "push"
-      when: never
-    - if: $CI_PIPELINE_SOURCE == "merge_request_event"
-    - if: $CI_COMMIT_BRANCH && $CI_OPEN_MERGE_REQUESTS
-      when: never
-    - if: $CI_COMMIT_BRANCH
   extends: .test_template
   script:
     - python -m pytest --cov=FINE test/
-  artifacts:
-    paths:
-      - test/data/*
-    expire_in: 1 week
 
-test-formatting:
-  #  Switch from branch pipeline to merge pipeline once a merge request has been
-  #  created on the branch.
-  rules:
-    - if: $CI_COMMIT_BRANCH && $CI_OPEN_MERGE_REQUESTS && $CI_PIPELINE_SOURCE == "push"
-      when: never
-    - if: $CI_PIPELINE_SOURCE == "merge_request_event"
-    - if: $CI_COMMIT_BRANCH && $CI_OPEN_MERGE_REQUESTS
-      when: never
[…]
