+++ requirements.yml
@@ -25,7 +25,7 @@ dependencies:
   - conda-forge::rasterio
   - conda-forge::netcdf4
   - conda-forge::tsam
-  - pwlf
+  - conda-forge::pwlf
 
   - pip:
       - .
