+++ .gitlab-ci.yml
@@ -43,11 +43,13 @@ variables:
       when: never
     - if: $CI_COMMIT_BRANCH && $CI_OPEN_MERGE_REQUESTS
       when: never
-    - changes:
+    - if: true
+      changes:
         - pyproject.toml
         - requirements.yml
         - requirements_dev.yml
       when: never
+    - when: on_success
 
 
 .test_docker_template:
