+++ .gitlab-ci.yml
@@ -43,8 +43,7 @@ variables:
       when: never
     - if: $CI_COMMIT_BRANCH && $CI_OPEN_MERGE_REQUESTS
       when: never
-    - if: true
-      changes:
+    - changes:
         - pyproject.toml
         - requirements.yml
         - requirements_dev.yml
