### 2023-11-16 778f468
Issue 41 add performance summary
