### 2023-11-20 551a556
Change order of build rules
