### 2023-11-22 9775515
Add Mamba background
