+++ requirements_dev.yml
@@ -3,7 +3,7 @@ name: FINE_dev
 dependencies:
   - conda-forge::python>=3.10
   - conda-forge::libiconv
-  - conda-forge::gdal=3.4
+  - conda-forge::gdal=3.4.3
   - conda-forge::geopandas
   - conda-forge::jupyter
   - conda-forge::notebook
@@ -33,7 +33,7 @@ dependencies:
   - conda-forge::black
   - conda-forge::pwlf
   - conda-forge::psutil
+  - conda-forge::gurobi-logtools
   
   - pip:
-      - grblogtools
       - -e .
