### 2023-12-05 e9c671d
Add frozen requirements to CI
