### 2023-12-05 1a8938a
Move requirements freeze to build stage
