+++ .gitlab-ci.yml
@@ -136,6 +136,7 @@ test-code-push-cache:
 
 create-frozen-requirements:
   extends: .test_template
+  stage: build
   script:
     - micromamba env export --no-md5 --no-build -n base > requirements_frozen.yaml
     # Replace `name: base` with `name: fine_frozen`
