+++ test/IOManagement/test_standardIO.py
@@ -1,2 +1,2 @@
-import FINE.IOManagement.standardIO as stdIO
+import fine.IOManagement.standardIO as stdIO
 import os
