+++ fine/utils.py
@@ -4,7 +4,7 @@ import warnings
 import numpy as np
 import pandas as pd
 
-import FINE as fn
+import fine as fn
 
 
 def isString(string):
