+++ requirements_frozen.yml
@@ -15,7 +15,7 @@ dependencies:
 - async-lru=2.0.4
 - attr=2.5.1
 - attrs=23.1.0
-- babel=2.13.1
+- babel=2.14.0
 - beautifulsoup4=4.12.2
 - black=23.11.0
 - bleach=6.1.0
@@ -42,7 +42,7 @@ dependencies:
 - colorama=0.4.6
 - comm=0.1.4
 - contourpy=1.2.0
-- coverage=7.3.2
+- coverage=7.3.3
 - curl=7.86.0
 - cycler=0.12.1
 - dbus=1.13.6
@@ -77,8 +77,8 @@ dependencies:
 - geotiff=1.7.1
 - gettext=0.21.1
 - giflib=5.2.1
-- glib=2.78.2
-- glib-tools=2.78.2
+- glib=2.78.3
+- glib-tools=2.78.3
 - glpk=5.0
 - gmp=6.3.0
 - gst-plugins-base=1.18.5
@@ -114,8 +114,8 @@ dependencies:
 - jupyter_console=6.6.3
 - jupyter_core=5.5.0
 - jupyter_events=0.9.0
-- jupyter_server=2.12.0
-- jupyter_server_terminals=0.4.4
+- jupyter_server=2.12.1
+- jupyter_server_terminals=0.5.0
 - jupyterlab=4.0.9
 - jupyterlab_pygments=0.3.0
 - jupyterlab_server=2.25.2
@@ -149,7 +149,7 @@ dependencies:
 - libgdal=3.4.3
 - libgfortran-ng=13.2.0
 - libgfortran5=13.2.0
-- libglib=2.78.2
+- libglib=2.78.3
 - libgomp=13.2.0
 - libiconv=1.17
 - libkml=1.3.0
@@ -219,11 +219,11 @@ dependencies:
 - openssl=1.1.1w
 - overrides=7.4.0
 - packaging=23.2
-- pandas=2.1.3
+- pandas=2.1.4
 - pandoc=3.1.3
 - pandocfilters=1.5.0
 - parso=0.8.3
-- pathspec=0.11.2
+- pathspec=0.12.1
 - pcre=8.45
 - pcre2=10.42
 - pexpect=4.8.0
@@ -241,8 +241,8 @@ dependencies:
 - postgresql=14.5
 - proj=9.0.0
 - prometheus_client=0.19.0
-- prompt-toolkit=3.0.41
-- prompt_toolkit=3.0.41
+- prompt-toolkit=3.0.42
+- prompt_toolkit=3.0.42
 - psutil=5.9.5
 - pthread-stubs=0.4
 - ptyprocess=0.7.0
@@ -274,7 +274,7 @@ dependencies:
 - qtpy=2.4.1
 - rasterio=1.2.10
 - readline=8.2
-- referencing=0.31.1
+- referencing=0.32.0
 - requests=2.31.0
 - rfc3339-validator=0.1.4
 - rfc3986-validator=0.1.1
@@ -316,8 +316,8 @@ dependencies:
 - traitlets=5.14.0
 - tsam=2.1.1
 - types-python-dateutil=2.8.19.14
-- typing-extensions=4.8.0
-- typing_extensions=4.8.0
+- typing-extensions=4.9.0
+- typing_extensions=4.9.0
 - typing_utils=0.1.0
 - tzcode=2023c
 - tzdata=2023c
@@ -331,7 +331,7 @@ dependencies:
 - websocket-client=1.7.0
 - wheel=0.42.0
 - widgetsnbextension=4.0.9
-- xarray=2023.11.0
+- xarray=2023.12.0
 - xerces-c=3.2.3
 - xlrd=2.0.1
 - xlsxwriter=3.1.9
