### 2024-01-09 2620e08
Update news entry
