+++ requirements_frozen.yml
@@ -1,356 +1,0 @@
-name: fine_frozen
-channels:
-- conda-forge
-dependencies:
-- _libgcc_mutex=0.1
-- _openmp_mutex=4.5
-- affine=2.4.0
-- alabaster=0.7.13
-- alsa-lib=1.2.3.2
-- anyio=4.1.0
-- argon2-cffi=23.1.0
-- argon2-cffi-bindings=21.2.0
-- arrow=1.3.0
-- asttokens=2.4.1
-- async-lru=2.0.4
-- attr=2.5.1
-- attrs=23.1.0
-- babel=2.14.0
-- beautifulsoup4=4.12.2
-- black=23.11.0
-- bleach=6.1.0
-- blosc=1.21.5
-- boost-cpp=1.74.0
-- branca=0.7.0
-- brotli=1.1.0
-- brotli-bin=1.1.0
-- brotli-python=1.1.0
-- bzip2=1.0.8
-- c-ares=1.23.0
-- ca-certificates=2023.11.17
-- cached-property=1.5.2
-- cached_property=1.5.2
-- cairo=1.16.0
-- certifi=2023.11.17
-- cffi=1.16.0
-- cfitsio=4.1.0
-- cftime=1.6.3
-- charset-normalizer=3.3.2
-- click=8.1.7
-- click-plugins=1.1.1
-- cligj=0.7.2
-- colorama=0.4.6
-- comm=0.1.4
-- contourpy=1.2.0
-- coverage=7.3.3
-- curl=7.86.0
-- cycler=0.12.1
-- dbus=1.13.6
-- debugpy=1.8.0
-- decorator=5.1.1
-- defusedxml=0.7.1
-- descartes=1.1.0
-- docutils=0.20.1
-- entrypoints=0.4
-- et_xmlfile=1.1.0
-- exceptiongroup=1.2.0
-- executing=2.0.1
-- expat=2.5.0
-- fftw=3.3.10
-- fiona=1.8.21
-- folium=0.15.1
-- font-ttf-dejavu-sans-mono=2.37
-- font-ttf-inconsolata=3.000
-- font-ttf-source-code-pro=2.038
-- font-ttf-ubuntu=0.83
-- fontconfig=2.14.2
-- fonts-conda-ecosystem=1
-- fonts-conda-forge=1
-- fonttools=4.46.0
-- fqdn=1.5.1
-- freetype=2.12.1
-- freexl=1.0.6
-- gdal=3.4.3
-- geopandas=0.14.1
-- geopandas-base=0.14.1
-- geos=3.10.2
-- geotiff=1.7.1
-- gettext=0.21.1
-- giflib=5.2.1
-- glib=2.78.3
-- glib-tools=2.78.3
-- glpk=5.0
-- gmp=6.3.0
-- gst-plugins-base=1.18.5
-- gstreamer=1.20.3
-- gurobi-logtools=3.0.0
-- hdf4=4.2.15
-- hdf5=1.12.1
-- icu=69.1
-- idna=3.6
-- imagesize=1.4.1
-- importlib-metadata=7.0.0
-- importlib_metadata=7.0.0
-- importlib_resources=6.1.1
-- iniconfig=2.0.0
-- ipykernel=6.26.0
-- ipython=8.18.1
-- ipywidgets=8.1.1
-- isoduration=20.11.0
-- jack=1.9.18
-- jedi=0.19.1
-- jinja2=3.1.2
-- joblib=1.3.2
-- jpeg=9e
-- json-c=0.16
-- json5=0.9.14
-- jsonpointer=2.4
-- jsonschema=4.20.0
[…]
