### 2024-01-15 f4410ea
312 add operation rate min
