+++ fine/component.py
@@ -650,18 +650,13 @@ class Component(metaclass=ABCMeta):
             esM.componentModelingDict.update({mdl: self.modelingClass()})
         esM.componentModelingDict[mdl].componentsDict.update({self.name: self})
 
-    def prepareTSAInput(
-        self, rateFix, rateMax, rateName, rateWeight, weightDict, data, ip
-    ):
+    def prepareTSAInput(self, rate, rateName, rateWeight, weightDict, data, ip):
         """
         Format the time series data of a component to fit the requirements of the time series aggregation package and
         return a list of formatted data.
 
-        :param rateFix: a fixed operation time series or None
-        :type rateFix: Pandas DataFrame or None
-
-        :param rateMax: a maximum operation time series or None
-        :type rateMax: Pandas DataFrame of None
+        :param rate: a fixed/maximum/minimum operation time series or None
+        :type rate: Pandas DataFrame or None
 
         :param rateName: name of the time series (to ensure uniqueness if a component has multiple relevant time series)
         :type rateName: string
@@ -681,17 +676,13 @@ class Component(metaclass=ABCMeta):
         :return: data
         :rtype: Pandas DataFrame
         """
-        # rateFix/rateMax can be passed as a dict with investment periods
-        if isinstance(rateFix, dict):
-            rateFix = rateFix[ip]
-        else:
-            pass
-        if isinstance(rateMax, dict):
-            rateMax = rateMax[ip]
+        # rate can be passed as a dict with investment periods
+        if isinstance(rate, dict):
+            rate = rate[ip]
         else:
             pass
 
-        data_ = rateFix if rateFix is not None else rateMax
+        data_ = rate
         if data_ is not None:
             data_ = data_.copy()
             uniqueIdentifiers = [self.name + rateName + loc for loc in data_.columns]
@@ -1163,6 +1154,28 @@ class ComponentModel(metaclass=ABCMeta):
             pyomo.Set(dimen=3, initialize=declareOpConstrSet3),
         )
 
+    def declareOpConstrSet4(self, pyM, constrSetName, rateMin):
+        """
+        Declare set of locations and components for which  hasCapacityVariable is set to True and a minimum
+        operation rate is given.
+        """
+        compDict, abbrvName = self.componentsDict, self.abbrvName
+        varSet = getattr(pyM, "operationVarSet_" + abbrvName)
+
+        def declareOpConstrSet4(pyM):
+            return (
+                (loc, compName, ip)
+                for loc, compName, ip in varSet
+                if compDict[compName].hasCapacityVariable
+                and getattr(compDict[compName], rateMin) is not None
+            )
+
+        setattr(
+            pyM,
+            constrSetName + "4_" + abbrvName,
+            pyomo.Set(dimen=3, initialize=declareOpConstrSet4),
+        )
+
     def declareOpConstrSetMinPartLoad(self, pyM, constrSetName):
         """
         Declare set of locations and components for which partLoadMin is not None.
@@ -1183,7 +1196,9 @@ class ComponentModel(metaclass=ABCMeta):
             pyomo.Set(dimen=3, initialize=declareOpConstrSetMinPartLoad),
         )
 
-    def declareOperationModeSets(self, pyM, constrSetName, rateMax, rateFix):
+    def declareOperationModeSets(
+        self, pyM, constrSetName, rateMax, rateFix, rateMin=None
+    ):
         """
         Declare operating mode sets.
 
@@ -1196,12 +1211,18 @@ class ComponentModel(metaclass=ABCMeta):
         :param rateMax: attribute of the considered component which stores the maximum operation rate data.
         :type rateMax: string
 
+        :param rateMax: attribute of the considered component which stores the minimum operation rate data.
+        :type rateMax: string
+
[…]
