### 2024-01-22 #5965 @t-gross
> 👋 @t-gross, thanks for your submission to JOSS. I'll be your handling editor for this submission. As I start to look for good reviewers for this submission, I'd appreciate any suggestions for reviewers you think might be a good fit here. Thank you!

Thank you for taking over as editor. As suggestions for reviewers, we could imagine leonardgoeke being a good candidate. From searching in the list of people that have already agreed to review, olejandro and GregorBecker might fit.
