### 2024-01-22 #5965 @arfon
:wave: @olejandro @GregorBecker @leonardgoeke @KriNiTi – would any of you be willing to review this submission for JOSS? The submission under consideration is _ETHOS.FINE: A Framework for Integrated Energy System Assessment_

Many thanks
Arfon
