### 2024-01-23 ec08a4d
Change affinity to metric
