### 2024-01-24 #6274 @editorialbot
[REVIEW]: ETHOS.FINE: A Framework for Integrated Energy System Assessment
**Submitting author:** @t-gross (<a href="http://orcid.org/0000-0003-1792-5598">Theresa Gross</a>)
**Repository:** https://github.com/FZJ-IEK3-VSA/FINE
**Branch with paper.md** (empty if default branch): paper
**Version:** v2.3.0
