### 2024-02-07 8dbed2a
Add channel priority, allow all gdal 3.4
