+++ requirements.yml
@@ -1,33 +1,36 @@
 name: fine
+
 channels:
+  - conda-forge
   - nodefaults
+
 dependencies:
-  - conda-forge::python>=3.10
-  - conda-forge::libiconv
-  - conda-forge::gdal=3.4.3
-  - conda-forge::geopandas
-  - conda-forge::jupyter
-  - conda-forge::notebook
-  - conda-forge::libpysal
-  - conda-forge::pip
-  - conda-forge::descartes
-  - conda-forge::glpk
-  - conda-forge::openpyxl
-  - conda-forge::matplotlib
-  - conda-forge::xlrd
-  - conda-forge::pyomo
-  - conda-forge::numpy
-  - conda-forge::pandas
-  - conda-forge::networkx
-  - conda-forge::scipy
-  - conda-forge::scikit-learn>=1.2
-  - conda-forge::xarray
-  - conda-forge::rasterio
-  - conda-forge::netcdf4
-  - conda-forge::tsam
-  - conda-forge::pwlf
-  - conda-forge::psutil
-  - conda-forge::gurobi-logtools
+  - python>=3.10
+  - libiconv
+  - gdal=3.4.*
+  - geopandas
+  - jupyter
+  - notebook
+  - libpysal
+  - pip
+  - descartes
+  - glpk
+  - openpyxl
+  - matplotlib
+  - xlrd
+  - pyomo
+  - numpy
+  - pandas
+  - networkx
+  - scipy
+  - scikit-learn>=1.2
+  - xarray
+  - rasterio
+  - netcdf4
+  - tsam
+  - pwlf
+  - psutil
+  - gurobi-logtools
 
   - pip:
       - .
