### 2024-02-07 e36118b
Fix Python 3.10, remove gdal
