+++ requirements_dev.yml
@@ -5,9 +5,8 @@ channels:
   - nodefaults
 
 dependencies:
-  - python>=3.10
+  - python=3.10
   - libiconv
-  - gdal=3.4.*
   - geopandas
   - jupyter
   - notebook
@@ -17,7 +16,6 @@ dependencies:
   - glpk
   - openpyxl
   - matplotlib
-  - pillow
   - xlrd
   - pyomo
   - numpy
@@ -29,6 +27,10 @@ dependencies:
   - rasterio
   - netcdf4
   - tsam
+  - pwlf
+  - psutil
+  - gurobi-logtools
+  # develop
   - sphinx
   - sphinx_rtd_theme
   - myst-parser
@@ -36,9 +38,6 @@ dependencies:
   - pytest-cov
   - nbval
   - black
-  - pwlf
-  - psutil
-  - gurobi-logtools
-  
+
   - pip:
       - -e .
