+++ requirements.yml
@@ -5,9 +5,8 @@ channels:
   - nodefaults
 
 dependencies:
-  - python>=3.10
+  - python=3.10
   - libiconv
-  - gdal=3.4.*
   - geopandas
   - jupyter
   - notebook
