### 2024-02-07 f5c7f42
Use workstations for ci
