+++ .gitlab-ci.yml
@@ -16,8 +16,6 @@ variables:
 
 .test_template:
   stage: test
-  tags:
-    - workstation
   variables:
     CONDA_PKGS_DIRS: "$CI_PROJECT_DIR/.cache/pkgs"
   cache:
@@ -40,6 +38,7 @@ variables:
       when: never
     - if: $CI_COMMIT_BRANCH && $CI_OPEN_MERGE_REQUESTS
       when: never
+  retry: 1
 
 
 .test_docker_template_noupdate:
@@ -109,8 +108,6 @@ variables:
 
 .build_template:
   stage: build
-  tags:
-    - workstation
   image: docker@sha256:c8bb6fa5388b56304dd770c4bc0478de81ce18540173b1a589178c0d31bfce90
   services:
     - docker:dind@sha256:c8bb6fa5388b56304dd770c4bc0478de81ce18540173b1a589178c0d31bfce90
