### 2024-02-08 f155bb9
Add test workflow for github
