+++ .github/workflows/test.yml
@@ -79,7 +79,7 @@ jobs:
           mamba env create --name fine_env --yes --file requirements_dev.yml
           conda run --name fine_env pip install . --no-deps
           echo "Installation done"
-          conda list
+          conda list --name fine_env
           echo "libaries printed"
           echo "start pytest"
           conda run --name fine_env pytest
