### 2024-02-13 1be20b4
Resolve "test_perfectForesight_capacityBounds needs refactoring"
