### 2024-02-14 FZJ-IEK3-VSA/FINE#47 @GregorBecker
[JOSS] Installation 
Hi this is a question regarding the installation from [JOSS Review 6274](https://github.com/openjournals/joss-reviews/issues/6274)

I'm not quite sure why you decided to use a mamba-only installation procedure of your energy system modeling tool.

Since I am developing without the use of conda, I had to install micromamba exclusively for your package "FINE".

It would be better to give users the flexibility to choose their virtual environment on their own, so that they can either use Conda, mircomamba or a virtual environment solution of their choice. It is necessary to mention the possibility of using a [Python virtual environment](https://docs.python.org/3/library/venv.html) as an alternative which does not require additional installations.

Possible solution would be (for MacOS users):

python3 -m venv _Your Directory_
pip3 install fine

In my opinion, a description of the creation of virtual environments would not be necessary. A link to the relevant instructions would be sufficient.

Additionally it would be nice if the installation instruction in your documentation gets it's own menu button to be easier accessable via the documentation.

Kind Regards
@GregorBecker
