### 2024-02-15 FZJ-IEK3-VSA/FINE#48 @GregorBecker
[JOSS] Gitlab contents
Hi this is a question regarding the issue templates and files related to gitlab from [JOSS Review 6274](https://github.com/openjournals/joss-reviews/issues/6274).

While quickly scanning your repository I've found a ".gitlab"-folder providing a templates for issue. Why don't you use them within your github-repository. Otherwise I would recommend removing them from the repository because they are not relevant for the presented software.

The same problem occurs with .gitlab-ci.yml.

Kind regards
@GregorBecker
