### 2024-02-15 FZJ-IEK3-VSA/FINE#50 @GregorBecker
[JOSS] Documentation - Community guidelines
Hi, this issue is about the Community guidelines within the [JOSS Review 6274](https://github.com/openjournals/joss-reviews/issues/6274).

Please add a CONTRIBUTING.md and a CODE_OF_CONDUCT.md to your repository that describes how to handle feature & bug issues and pull requests to reduce the hurdle for possible contributors. For this purpose, you can use tools such as [contributing-gen](https://bttger.github.io/contributing-gen-web/).

In combination with moving the templates provided in the .gitlab folder to a .github folder, these would fulfill the community guideline ([Issue 48](https://github.com/FZJ-IEK3-VSA/FINE/issues/48)).

Kind regards
@GregorBecker
