### 2024-02-19 #6274 @leonardgoeke
Overall, the introduction of ETHOS.FINE is clear, comprehensible, and relevant. I have a few points listed below, following the guidelines above.

### Functionality

The installation guide was clear, but mamba as a package installer was new to me and something I had to install separately. Therefore, it is worth briefly mentioning and explaining this choice in the paper. The long list of peer-reviewed publications applying the tool proves its functionality. Although this could be interesting in this context, there are no claims regarding performance.

### General checks

All good here.

### Documentation

The repository names the package's dependencies but only specifies the version in one case. In my experience with Python, packages are often not compatible with the newest version of a dependency. As a result, the user must invest a lot of time reconstructing a functioning configuration if the dependencies do not specify a version.

The documentation shows how to cite the package but gives no details on contributing or reporting problems.

### Software Paper

The current state-of-the-art and need for the software is not discussed in the paper. The one paper that is supposed to prove the need is outdated and the tool is not differentiated from others in the field (Calliope, GenX, PyPSA, AnyMOD, etc.).

Earlier, the paper mentioned complexity reduction as a key capability of FINE. Therefore, some indicative information on performance is sensible, for instance, how long it takes to generate and solve a model with a certain number of time-steps, regions, and technologies. In this context, I would also like to know if the tools use pyomo as a backend for historical reasons or if it has an advantage over Linopy in the specific case.
