### 2024-02-22 2624ca8
add functionality for componentLimit for sourceSinkModel
