### 2024-02-22 364f6d4
added test for componentLimit
