### 2024-02-22 c3fdac0
added processing of componentLimit
