### 2024-02-22 805c486
dont build constraintif componentLimit is None
