+++ test/IOManagement/test_dictIO.py
@@ -23,6 +23,8 @@ def test_export_to_dict_minimal(minimal_test_esM):
                 "balanceLimit",
                 "pathwayBalanceLimit",
                 "annuityPerpetuity",
+                "componentLimitEligibility",
+                "componentLimit",
             ),
             (
                 minimal_test_esM.locations,
@@ -40,6 +42,8 @@ def test_export_to_dict_minimal(minimal_test_esM):
                 minimal_test_esM.balanceLimit,
                 minimal_test_esM.pathwayBalanceLimit,
                 minimal_test_esM.annuityPerpetuity,
+                minimal_test_esM.componentLimitEligibility,
+                minimal_test_esM.componentLimit,
             ),
         )
     )
@@ -98,6 +102,8 @@ def test_export_to_dict_singlenode(single_node_test_esM):
                 "balanceLimit",
                 "pathwayBalanceLimit",
                 "annuityPerpetuity",
+                "componentLimitEligibility",
+                "componentLimit",
             ),
             (
                 single_node_test_esM.locations,
@@ -115,6 +121,8 @@ def test_export_to_dict_singlenode(single_node_test_esM):
                 single_node_test_esM.balanceLimit,
                 single_node_test_esM.pathwayBalanceLimit,
                 single_node_test_esM.annuityPerpetuity,
+                single_node_test_esM.componentLimitEligibility,
+                single_node_test_esM.componentLimit,
             ),
         )
     )
@@ -175,6 +183,8 @@ def test_export_to_dict_multinode(multi_node_test_esM_init):
                 "balanceLimit",
                 "pathwayBalanceLimit",
                 "annuityPerpetuity",
+                "componentLimitEligibility",
+                "componentLimit",
             ),
             (
                 multi_node_test_esM_init.locations,
@@ -192,6 +202,8 @@ def test_export_to_dict_multinode(multi_node_test_esM_init):
                 multi_node_test_esM_init.balanceLimit,
                 multi_node_test_esM_init.pathwayBalanceLimit,
                 multi_node_test_esM_init.annuityPerpetuity,
+                multi_node_test_esM_init.componentLimitEligibility,
+                multi_node_test_esM_init.componentLimit,
             ),
         )
     )
