### 2024-03-04 f8b5f6f
Hotfix for nbval
Pytest 8.1 breaks nbval therefor we limit pytest to <8.1 until the issue is resolved.
