### 2024-03-05 10c8713
Add commissioning Min, Max, Fix constraints
