+++ fine/subclasses/lopf.py
@@ -33,6 +33,9 @@ class LinearOptimalPowerFlow(Transmission):
         partLoadMin=None,
         sharedPotentialID=None,
         capacityFix=None,
+        commissioningMin=None,
+        commissioningMax=None,
+        commissioningFix=None,
         isBuiltFix=None,
         investPerCapacity=0,
         investIfBuilt=0,
@@ -79,6 +82,9 @@ class LinearOptimalPowerFlow(Transmission):
             partLoadMin=partLoadMin,
             sharedPotentialID=sharedPotentialID,
             capacityFix=capacityFix,
+            commissioningMin=commissioningMin,
+            commissioningMax=commissioningMax,
+            commissioningFix=commissioningFix,
             isBuiltFix=isBuiltFix,
             investPerCapacity=investPerCapacity,
             investIfBuilt=investIfBuilt,
