+++ fine/utils.py
@@ -545,105 +545,92 @@ def checkLocationSpecficDesignInputParams(comp, esM):
             raise ValueError('QPcostScale must ba a number between "0" and "1".')
 
 
-def checkAndSetCapacityBounds(esM, name, capacityMin, capacityMax, capacityFix):
-    checkInvestmentPeriodParameters(name, capacityMin, esM.investmentPeriodNames)
-    checkInvestmentPeriodParameters(name, capacityMax, esM.investmentPeriodNames)
-    checkInvestmentPeriodParameters(name, capacityFix, esM.investmentPeriodNames)
-
-    def _checkAndSet(name, param):
+def processBoundParams(esM, param):
+    processedParam = {}
+    for ip in esM.investmentPeriods:
+        if param is None:
+            processedParam[ip] = None
         if isinstance(param, dict):
-            if all(x is None for x in param.values()):
-                return None
-            if any(x is None for x in param.values()):
-                raise ValueError(
-                    f"{name} cannot switch between None and values for different investment periods."
-                )
-
-        processedParam = {}
-        for ip in esM.investmentPeriods:
-            if param is None:
+            if param[esM.investmentPeriodNames[ip]] is None:
                 processedParam[ip] = None
-            if isinstance(param, dict):
+            else:
                 processedParam[ip] = castToSeries(
                     param[esM.investmentPeriodNames[ip]], esM
                 )
-            elif isinstance(param, pd.DataFrame) or isinstance(param, pd.Series):
-                processedParam[ip] = castToSeries(param, esM)
-            elif isinstance(param, int) or isinstance(param, float):
-                processedParam[ip] = castToSeries(param, esM)
-        return processedParam
+        elif isinstance(param, pd.DataFrame) or isinstance(param, pd.Series):
+            processedParam[ip] = castToSeries(param, esM)
+        elif isinstance(param, int) or isinstance(param, float):
+            processedParam[ip] = castToSeries(param, esM)
+    return processedParam
+
+
+def checkForConsistency(paramName, valueName, name, capacityBound):
+    if isinstance(capacityBound, dict):
+        if any(x is not None for x in capacityBound.values()):
+            if not all(x is not None for x in capacityBound.values()):
+                raise ValueError(
+                    "A mix between None and specified values is not allowed "
+                    + f"between investment periods for {paramName}{valueName} for {name}."
+                )
+
+
+def checkAndSetBounds(esM, name, paramName, MinVal, MaxVal, FixVal):
+    checkInvestmentPeriodParameters(name, MinVal, esM.investmentPeriodNames)
+    checkInvestmentPeriodParameters(name, MaxVal, esM.investmentPeriodNames)
+    checkInvestmentPeriodParameters(name, FixVal, esM.investmentPeriodNames)
 
     # set up parameter as dict with investment periods as keys and
     # dataframe with locations as values
-    processedCapacityMin = _checkAndSet(name, capacityMin)
-    processedCapacityMax = _checkAndSet(name, capacityMax)
-    processedCapacityFix = _checkAndSet(name, capacityFix)
+    processedMinVal = processBoundParams(esM, MinVal)
+    processedMaxVal = processBoundParams(esM, MaxVal)
+    processedFixVal = processBoundParams(esM, FixVal)
 
     for ip in esM.investmentPeriods:
-        if (
-            processedCapacityMin[ip] is not None
-            and (processedCapacityMin[ip] < 0).any()
-        ):
-            raise ValueError("capacityMin values smaller than 0 were detected.")
+        if processedMinVal[ip] is not None and (processedMinVal[ip] < 0).any():
+            raise ValueError(
+                f"{paramName}Min values for {name} smaller than 0 were detected."
+            )
 
-        if (
-            processedCapacityFix[ip] is not None
[…]
