+++ fine/component.py
@@ -31,6 +31,9 @@ class Component(metaclass=ABCMeta):
         sharedPotentialID=None,
         linkedQuantityID=None,
         capacityFix=None,
+        commissioningMin=None,
+        commissioningMax=None,
+        commissioningFix=None,
         isBuiltFix=None,
         investPerCapacity=0,
         investIfBuilt=0,
@@ -177,7 +180,7 @@ class Component(metaclass=ABCMeta):
 
         :param sharedPotentialID: if specified, indicates that the component has to share its maximum
             potential capacity with other components (e.g. due to space limitations). The shares of how
-            much of the maximum potential is used have to add up to less then 100%.
+            much of the maximum potential is used have to add up to less than 100%.
             |br| * the default value is None
         :type sharedPotentialID: string
 
@@ -192,6 +195,58 @@ class Component(metaclass=ABCMeta):
             * If dimension=2dim, it has to be a Pandas Series or DataFrame.
             |br| * the default value is None
         :type capacityFix:
+            * None or
+            * float or
+            * int or
+            * Pandas Series with positive (>=0) values. The indices of the series have to equal the in the
+              energy system model specified locations (dimension=1dim) or connections between these locations
+              in the format of 'loc1' + '_' + 'loc2' (dimension=2dim) or
+            * Pandas DataFrame with positive (>=0) values. The row and column indices of the DataFrame have
+              to equal the in the energy system model specified locations. or
+            * Dict with investment periods as keys and one of the options above as values.
+
+        :param commissioningMin: if specified, indicates the minimum commissioning for the respective
+            investment period. The type of this parameter depends on the dimension of the component:
+            * If dimension=1dim, it has to be a Pandas Series.
+            * If dimension=2dim, it has to be a Pandas Series or DataFrame.
+            If binary decision variables are declared, commissioningMin is only used
+            if the component is built.
+            |br| * the default value is None
+        :type commissioningMin:
+
+            * None or
+            * float or
+            * int or
+            * Pandas Series with positive (>=0) values. The indices of the series have to equal the in the
+              energy system model specified locations (dimension=1dim) or connections between these locations
+              in the format of 'loc1' + '_' + 'loc2' (dimension=2dim) or
+            * Pandas DataFrame with positive (>=0) values. The row and column indices of the DataFrame have
+              to equal the in the energy system model specified locations. or
+            * Dict with investment periods as keys and one of the options above as values.
+
+        :param commissioningMax: if specified, indicates the maximum commissioning for the respective
+            investment period. The type of this parameter depends on the dimension of the component:
+            * If dimension=1dim, it has to be a Pandas Series.
+            * If dimension=2dim, it has to be a Pandas Series or DataFrame.
+            |br| * the default value is None
+        :type commissioningMax:
+
+            * None or
+            * float or
+            * int or
+            * Pandas Series with positive (>=0) values. The indices of the series have to equal the in the
+              energy system model specified locations (dimension=1dim) or connections between these locations
+              in the format of 'loc1' + '_' + 'loc2' (dimension=2dim) or
+            * Pandas DataFrame with positive (>=0) values. The row and column indices of the DataFrame have
+              to equal the in the energy system model specified locations. or
+            * Dict with investment periods as keys and one of the options above as values.
+
+        :param commissioningFix: if specified, indicates the fixed commissioning for the respective
[…]
