+++ fine/sourceSink.py
@@ -33,6 +33,9 @@ class Source(Component):
         sharedPotentialID=None,
         linkedQuantityID=None,
         capacityFix=None,
+        commissioningMin=None,
+        commissioningMax=None,
+        commissioningFix=None,
         isBuiltFix=None,
         investPerCapacity=0,
         investIfBuilt=0,
@@ -265,6 +268,9 @@ class Source(Component):
             sharedPotentialID=sharedPotentialID,
             linkedQuantityID=linkedQuantityID,
             capacityFix=capacityFix,
+            commissioningMin=commissioningMin,
+            commissioningMax=commissioningMax,
+            commissioningFix=commissioningFix,
             isBuiltFix=isBuiltFix,
             investPerCapacity=investPerCapacity,
             investIfBuilt=investIfBuilt,
@@ -615,6 +621,9 @@ class Sink(Source):
         sharedPotentialID=None,
         linkedQuantityID=None,
         capacityFix=None,
+        commissioningMin=None,
+        commissioningMax=None,
+        commissioningFix=None,
         isBuiltFix=None,
         investPerCapacity=0,
         investIfBuilt=0,
@@ -664,6 +673,9 @@ class Sink(Source):
             sharedPotentialID=sharedPotentialID,
             linkedQuantityID=linkedQuantityID,
             capacityFix=capacityFix,
+            commissioningMin=commissioningMin,
+            commissioningMax=commissioningMax,
+            commissioningFix=commissioningFix,
             isBuiltFix=isBuiltFix,
             investPerCapacity=investPerCapacity,
             investIfBuilt=investIfBuilt,
