### 2024-03-06 573c26d
fix wrong dimension in checkAndSetInvestmentPeriodTimeSeries
