### 2024-03-06 dfadaae
add componentLimit for Storage
