+++ fine/storage.py
@@ -57,6 +57,7 @@ class Storage(Component):
         floorTechnicalLifetime=True,
         socOffsetDown=-1,
         socOffsetUp=-1,
+        componentLimitID=None,
         stockCommissioning=None,
     ):
         """
@@ -288,6 +289,7 @@ class Storage(Component):
         self.socOffsetUp = socOffsetUp
         self.socOffsetDown = socOffsetDown
         self.modelingClass = StorageModel
+        self.componentLimitID = componentLimitID
 
         # opexPerChargeOperation
         self.opexPerChargeOperation = opexPerChargeOperation
@@ -1653,6 +1655,66 @@ class StorageModel(ComponentModel):
                 for comp in self.componentsDict.values()
             ]
         )
+    
+    def getComponentLimitContribution(
+        self, esM, pyM, timeSeriesAggregation, ip, loc, componentNames, type
+    ):
+        """
+        Get contribution to componentLimitConstraint (Further read in EnergySystemModel).
+        Sum of the operation time series of a SourceSink component is used as the componentLimit contribution:
+
+        - If component is a Source it contributes with a positive sign to the limit. Example: Electricity Purchase
+        - A Sink contributes with a negative sign. Example: Sale of electricity
+
+        :param esM: EnergySystemModel instance representing the energy system in which the component should be modeled.
+        :type esM: esM - EnergySystemModel class instance
+
+        :param pym: pyomo ConcreteModel which stores the mathematical formulation of the model.
+        :type pym: pyomo ConcreteModel
+
+        :param timeSeriesAggregation: states if the optimization of the energy system model should be done with
+
+            (a) the full time series (False) or
+            (b) clustered time series data (True).
+
+        :type timeSeriesAggregation: boolean
+
+        :param ip: investment period of transformation path analysis.
+        :type ip: int
+
+        :param ID: ID of the regarded componentLimitConstraint
+        :param ID: string
+
+        :param loc: Name of the regarded location (locations are defined in the EnergySystemModel instance)
+        :type loc: string
+
+        :param componentNames: Names of components which contribute to the component limit
+        :type componentNames: list
+
+        :param type: Type of the variable ("operation" or "capacity")
+        :type type: string
+
+        """
+        compDict, abbrvName = self.componentsDict, self.abbrvName
+        if type == "operation":
+            raise NotImplementedError("ComponentLimit Constraint for operation not implemented for Storage components.")
+        elif type == "capacity":
+            capVar = getattr(pyM, "cap_" + abbrvName)
+            balance = sum(
+                capVar[loc, compName, ip]
+                for compName in compDict.keys()
+                if compName in componentNames
+                and compDict[compName].processedLocationalEligibility[loc] == 1
+            )
+            if isinstance(balance, int) or isinstance(balance, float):
+                return None
+            else:
+                return balance
+        else:
+            raise ValueError(
+                "Invalid type in ComponentLimit Contraint. Please choose 'operation' or 'capacity'."
+            )
+
 
     def getCommodityBalanceContribution(self, pyM, commod, loc, ip, p, t):
         """Get contribution to a commodity balance.
