### 2024-03-06 385a54e
First implementation steps for including multi-objective-optimization into FINE. Running the current state requires a customized version of the package pyaugmecon - therefore this branch might not work for you (yet)!
