+++ FINE/storage.py
@@ -58,6 +58,7 @@ class Storage(Component):
         socOffsetDown=-1,
         socOffsetUp=-1,
         stockCommissioning=None,
+        materialDemandPerCapacity=None,
     ):
         """
         Constructor for creating an Storage class instance.
@@ -265,6 +266,7 @@ class Storage(Component):
             technicalLifetime=technicalLifetime,
             stockCommissioning=stockCommissioning,
             floorTechnicalLifetime=floorTechnicalLifetime,
+            materialDemandPerCapacity=materialDemandPerCapacity,
         )
 
         # Set general storage component data: chargeRate, dischargeRate, chargeEfficiency, dischargeEfficiency,
@@ -1669,7 +1671,7 @@ class StorageModel(ComponentModel):
             if commod == self.componentsDict[compName].commodity
         )
 
-    def getObjectiveFunctionContribution(self, esM, pyM):
+    def getObjectiveFunctionContribution(self, esM, pyM, objective="costs"):
         """
         Get contribution to the objective function.
 
@@ -1679,77 +1681,83 @@ class StorageModel(ComponentModel):
         :param pyM: pyomo ConcreteModel which stores the mathematical formulation of the model.
         :type pyM: pyomo ConcreteModel
         """
-        compDict, abbrvName = self.componentsDict, self.abbrvName
+        if objective not in ["costs", "material_demand"]:
+            raise NotImplementedError("The chosen objective is not supported yet.")
 
-        capexCap = self.getEconomicsDesign(
-            pyM,
-            esM,
-            ["processedInvestPerCapacity"],
-            lifetimeAttr="ipEconomicLifetime",
-            varName="commis",
-            divisorName="CCF",
-        )
-        capexDec = self.getEconomicsDesign(
-            pyM,
-            esM,
-            ["processedInvestIfBuilt"],
-            lifetimeAttr="ipEconomicLifetime",
-            varName="commisBin",
-            divisorName="CCF",
-        )
-        opexCap = self.getEconomicsDesign(
-            pyM,
-            esM,
-            ["processedOpexPerCapacity"],
-            lifetimeAttr="ipTechnicalLifetime",
-            varName="commis",
-        )
-        opexDec = self.getEconomicsDesign(
-            pyM,
-            esM,
-            ["processedOpexIfBuilt"],
-            lifetimeAttr="ipTechnicalLifetime",
-            varName="commisBin",
-        )
-        opexOp1 = self.getEconomicsOperation(
-            pyM,
-            esM,
-            "TD",
-            ["processedOpexPerChargeOperation"],
-            "chargeOp",
-            "operationVarDict",
-        )
-        opexOp2 = self.getEconomicsOperation(
-            pyM,
-            esM,
-            "TD",
-            ["processedOpexPerDischargeOperation"],
-            "dischargeOp",
-            "operationVarDict",
-        )
+        if objective == "costs":
+            compDict, abbrvName = self.componentsDict, self.abbrvName
 
-        offsetUp = getattr(pyM, "stateOfChargeOffsetUp_" + abbrvName)
-        offsetDown = getattr(pyM, "stateOfChargeOffsetDown_" + abbrvName)
+            capexCap = self.getEconomicsDesign(
+                pyM,
+                esM,
+                ["processedInvestPerCapacity"],
+                lifetimeAttr="ipEconomicLifetime",
+                varName="commis",
+                divisorName="CCF",
+            )
+            capexDec = self.getEconomicsDesign(
+                pyM,
+                esM,
+                ["processedInvestIfBuilt"],
+                lifetimeAttr="ipEconomicLifetime",
+                varName="commisBin",
+                divisorName="CCF",
+            )
+            opexCap = self.getEconomicsDesign(
[…]
