### 2024-03-07 9ec53e3
Add documentation for the input parameters and Jupyter notebook example
