### 2024-03-07 cac50a8
Resolve "Use linting with ruff"
