+++ test/perfectForesight/test_perfectForesight_netPresentValue.py
@@ -1,6 +1,4 @@
-import fine as fn
 import numpy as np
-import pytest
 
 
 def test_perfectForesight_netPresentValue(perfectForesight_test_esM):
