+++ fine/IOManagement/utilsIO.py
@@ -1,4 +1,3 @@
-import math
 import numpy as np
 import pandas as pd
 import xarray as xr
