+++ fine/storage.py
@@ -460,7 +460,7 @@ class Storage(Component):
 
         """
         weightDict, data = {}, []
-        I = [
+        tsa_input = [
             (
                 self.fullChargeOpRateFix,
                 self.fullChargeOpRateMax,
@@ -475,7 +475,7 @@ class Storage(Component):
             ),
         ]
 
-        for rateFix, rateMax, rateName, rateWeight in I:
+        for rateFix, rateMax, rateName, rateWeight in tsa_input:
             if rateFix:
                 weightDict, data = self.prepareTSAInput(
                     rateFix, rateName, rateWeight, weightDict, data, ip
@@ -1668,7 +1668,7 @@ class StorageModel(ComponentModel):
             \\text{C}^{comp,comm}_{loc,ip,p,t} = op^{comp,discharge}_{loc,ip,p,t} - op^{comp,charge}_{loc,ip,p,t}
 
         """
-        compDict, abbrvName = self.componentsDict, self.abbrvName
+        abbrvName = self.abbrvName
         chargeOp, dischargeOp = (
             getattr(pyM, "chargeOp_" + abbrvName),
             getattr(pyM, "dischargeOp_" + abbrvName),
