### 2024-03-07 0f76768
Add publishing to pipeline
