+++ .gitlab-ci.yml
@@ -3,6 +3,7 @@ image: mambaorg/micromamba
 stages:
   - test
   - build
+  - publish
 
 variables:
   DOCKER_HOST: tcp://docker:2375
@@ -228,3 +229,33 @@ build-dev:
     - docker push jugit-registry.fz-juelich.de/iek-3/shared-code/fine/fine-dev:latest
   rules:
     - if: $CI_COMMIT_BRANCH == "develop"
+
+get-current-version:
+  stage: publish
+  script:
+    - pip install setuptools_scm
+    - echo "This version will be used to publish to pypi."
+    - python -m setuptools_scm
+  only:
+    - tags
+  # rules:
+  #   - if: $CI_COMMIT_TAG
+
+pypi-upload:
+  stage: publish
+  script:
+    - python -m pip install --upgrade build
+    - python -m pip install --upgrade twine
+    - python -m build
+    # Publish to testpypi
+    - python -m twine upload --repository testpypi -u __token__ -p $PYPI_TOKEN dist/*
+    # Test installation
+    - micromamba install -n base -y --file=requirements.yml
+    - python -m pip install --index-url https://test.pypi.org/simple/ --no-deps fine
+    - python -m pytest -n auto test/
+    # Publish to pypi 
+    - python -m twine upload -u __token__ -p $PYPI_TOKEN  dist/*
+  rules:
+    - if: $CI_COMMIT_TAG
+  when: manual
+
