+++ .gitlab-ci.yml
@@ -236,8 +236,6 @@ get-current-version:
     - pip install setuptools_scm
     - echo "This version will be used to publish to pypi."
     - python -m setuptools_scm
-  only:
-    - tags
   # rules:
   #   - if: $CI_COMMIT_TAG
 
