+++ test/IOManagement/test_standardIO.py
@@ -1,2 +1,0 @@
-import fine.IOManagement.standardIO as stdIO
-import os
