+++ fine/__init__.py
@@ -4,6 +4,8 @@ Last edited: February 06, 2020
 |br| @author: FINE Developer Team (FZJ IEK-3)
 """
 
+# ruff: noqa
+
 from .energySystemModel import EnergySystemModel
 from .sourceSink import Source, Sink
 from .conversion import Conversion
