+++ requirements_dev.yml
@@ -34,8 +34,8 @@ dependencies:
   - sphinx
   - sphinx_rtd_theme
   - myst-parser
-  - pytest<8.1
+  - pytest
   - pytest-cov
   - nbval
-  - black
   - pytest-xdist
+  - ruff

[…]
