### 2024-03-17 3d90ee9
make multiple balance limits per technology possible. fix issues with netdf in this regard. NOTE: currently not possible to create esM from xarray
