+++ fine/IOManagement/xarrayIO.py
@@ -367,7 +367,7 @@ def writeDatasetsToNetCDF(
                     _df = _df.reindex(sorted(_df.columns), axis=1)
                     for idx, row in _df.iterrows():
                         xarray_dataset.attrs.update(
-                            {f"{attr_name}.{idx}": row.to_numpy()}
+                            {f"{attr_name}.{idx}": row.to_list()}
                         )
 
                     # Delete the original attribute
