### 2024-03-17 9bfe7a7
fix add storage for ComponentBalanceLimit
