### 2024-03-17 e8fa8b1
fix bug for componentLimit Storage
