### 2024-03-18 FZJ-IEK3-VSA/FINE#48 @maurerle
As someone not related to this repository but with general development experience:
It is quite common to mirror a repository and have an internal version available on some other git server.
Due to the consequent way git archives files and puts hashes across the history of a file tree, it is not easily feasible to have this folder available in one VCS but not available in the other VCS.

The same goes for the `.gitlab-ci.yml` which might run on some mirror on some other git-server without relying on resources of GitHub.
So I do not think, that there is need to change this in any way.
