### 2024-03-20 65d3386
add componentLimitID attribute to Sink Class
