### 2024-03-21 3872af8
[JOSS, Issue #50] Documentation - Community Guidelines
