### 2024-03-21 FZJ-IEK3-VSA/FINE#50 @t-gross
Thank you for your feedback. We adjusted the section on how to contribute and added a CODE_OF_CONDUCT.md to our repository (currently available in develop-branch).

Currently, there are still some ongoing developments in an internal gitlab-version of this repository. We need to discuss how to deal with this. Nevertheless, thanks for your comment.
