### 2024-03-21 414d426
[JOSS Issue #49] Documentation - Example Usage
