+++ examples/NetCDF_to_save_and_set_up_model_instance/getModel.py
@@ -1,186 +1,0 @@
-import fine as fn
-import pandas as pd
-import numpy as np
-
-
-def getModel():
-    numberOfTimeSteps = 4
-    hoursPerTimeStep = 2190
-
-    # Create an energy system model instance
-    esM = fn.EnergySystemModel(
-        locations={"ElectrolyzerLocation", "IndustryLocation"},
-        commodities={"electricity", "hydrogen"},
-        numberOfTimeSteps=numberOfTimeSteps,
-        commodityUnitsDict={
-            "electricity": r"kW$_{el}$",
-            "hydrogen": r"kW$_{H_{2},LHV}$",
-        },
-        hoursPerTimeStep=hoursPerTimeStep,
-        costUnit="1 Euro",
-        lengthUnit="km",
-        verboseLogLevel=1,
-        balanceLimit=None,
-    )
-
-    # time step length [h]
-    timeStepLength = numberOfTimeSteps * hoursPerTimeStep
-
-    ### Buy electricity at the electricity market
-    costs = pd.DataFrame(
-        [
-            np.array(
-                [
-                    0.05,
-                    0.0,
-                    0.1,
-                    0.051,
-                ]
-            ),
-            np.array(
-                [
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                ]
-            ),
-        ],
-        index=["ElectrolyzerLocation", "IndustryLocation"],
-    ).T
-    revenues = pd.DataFrame(
-        [
-            np.array(
-                [
-                    0.0,
-                    0.01,
-                    0.0,
-                    0.0,
-                ]
-            ),
-            np.array(
-                [
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                ]
-            ),
-        ],
-        index=["ElectrolyzerLocation", "IndustryLocation"],
-    ).T
-    maxpurchase = (
-        pd.DataFrame(
-            [
-                np.array(
-                    [
-                        1e6,
-                        1e6,
-                        1e6,
-                        1e6,
-                    ]
-                ),
-                np.array(
-                    [
-                        0.0,
-                        0.0,
-                        0.0,
-                        0.0,
-                    ]
-                ),
-            ],
-            index=["ElectrolyzerLocation", "IndustryLocation"],
-        ).T
-        * hoursPerTimeStep
-    )
-    esM.add(
-        fn.Source(
-            esM=esM,
-            name="Electricity market",
-            commodity="electricity",
-            hasCapacityVariable=False,
-            operationRateMax=maxpurchase,
-            commodityCostTimeSeries=costs,
-            commodityRevenueTimeSeries=revenues,
-        )
-    )  # eur/kWh
-
-    ### Electrolyzers
-    esM.add(
-        fn.Conversion(
-            esM=esM,
-            name="Electrolyzers",
-            physicalUnit=r"kW$_{el}$",
-            commodityConversionFactors={"electricity": -1, "hydrogen": 0.7},
-            hasCapacityVariable=True,
-            investPerCapacity=500,  # euro/kW
-            opexPerCapacity=500 * 0.025,
-            interestRate=0.08,
-            economicLifetime=10,
-        )
-    )
-
-    ### Hydrogen filled somewhere
-    esM.add(
-        fn.Storage(
-            esM=esM,
-            name="Pressure tank",
-            commodity="hydrogen",
-            hasCapacityVariable=True,
-            capacityVariableDomain="continuous",
-            stateOfChargeMin=0.33,
-            investPerCapacity=0.5,  # eur/kWh
-            interestRate=0.08,
-            economicLifetime=30,
-        )
-    )
-
-    ### Hydrogen pipelines
-    esM.add(
-        fn.Transmission(
-            esM=esM,
-            name="Pipelines",
[…]
