+++ docs/source/tutorialDoc.rst
@@ -5,27 +5,27 @@ Tutorial
 The repository of ETHOS.FINE provides a certain amount of different `examples <https://github.com/FZJ-IEK3-VSA/FINE/tree/master/examples>` to get to know to the package. 
 The examples are sorted by their complexity and should provide a good overview on the scope of the package. 
 
-* `00_Tutorial <https://github.com/FZJ-IEK3-VSA/FINE/tree/master/examples/00_Tutorial>`
+* `00_Tutorial <https://github.com/FZJ-IEK3-VSA/FINE/tree/develop/examples/00_Tutorial>`
     * In this application, an energy supply system, consisting of two regions, is modeled and optimized. Recommended as starting point to get to know to ETHOS.FINE.
-* `01_1node_Energy_System_Workflow <https://github.com/FZJ-IEK3-VSA/FINE/tree/master/examples/01_1node_Energy_System_Workflow>`
+* `01_1node_Energy_System_Workflow <https://github.com/FZJ-IEK3-VSA/FINE/tree/develop/examples/01_1node_Energy_System_Workflow>`
     * In this application, a single region energy system is modeled and optimized. The system includes only a few technologies. 
-* `02_EnergyLand <https://github.com/FZJ-IEK3-VSA/FINE/tree/master/examples/02_EnergyLand>`
+* `02_EnergyLand <https://github.com/FZJ-IEK3-VSA/FINE/tree/develop/examples/02_EnergyLand>`
     * In this application, a single region energy system is modeled and optimized. Compared to the previous examples, this example includes a lot more technologies considered in the system. 
-* `03_Multi-regional_Energy_System_Workflow <https://github.com/FZJ-IEK3-VSA/FINE/tree/master/examples/03_Multi-regional_Energy_System_Workflow>`
+* `03_Multi-regional_Energy_System_Workflow <https://github.com/FZJ-IEK3-VSA/FINE/tree/develop/examples/03_Multi-regional_Energy_System_Workflow>`
     * In this application, an energy supply system, consisting of eight regions, is modeled and optimized. The example shows how to model multi-regional energy systems. The example also includes a notebook to get to know the optional performance summary. The summary shows how the optimization performed.
-* `04_Model_Run_from_Excel <https://github.com/FZJ-IEK3-VSA/FINE/tree/master/examples/04_Model_Run_from_Excel>`
+* `04_Model_Run_from_Excel <https://github.com/FZJ-IEK3-VSA/FINE/tree/develop/examples/04_Model_Run_from_Excel>`
     * ETHOS.FINE can also be run by excel. This example shows how to read and run a model using excel files.
-* `05_District_Optimization <https://github.com/FZJ-IEK3-VSA/FINE/tree/master/examples/05_District_Optimization>`
+* `05_District_Optimization <https://github.com/FZJ-IEK3-VSA/FINE/tree/develop/examples/05_District_Optimization>`
     * In this application, a small district is modeled and optimized. This example also includes binary decision variables.
-* `06_Water_Supply_System <https://github.com/FZJ-IEK3-VSA/FINE/tree/master/examples/06_Water_Supply_System>`
+* `06_Water_Supply_System <https://github.com/FZJ-IEK3-VSA/FINE/tree/develop/examples/06_Water_Supply_System>`
     * The application cases of ETHOS.FINE are not limited. This application shows how to model the water supply system. 
-* `07_NetCDF_to_save_and_set_up_model_instance <https://github.com/FZJ-IEK3-VSA/FINE/tree/master/examples/07_NetCDF_to_save_and_set_up_model_instance>`
+* `07_NetCDF_to_save_and_set_up_model_instance <https://github.com/FZJ-IEK3-VSA/FINE/tree/develop/examples/07_NetCDF_to_save_and_set_up_model_instance>`
[…]
