+++ .github/workflows/test.yml
@@ -78,7 +78,8 @@ jobs:
           ls
           echo "LS Done"
           cd fine
-          mamba create --name fine_env --yes --file requirements_dev.yml python=${{ matrix.python-version }}
+          mamba create --name fine_env --yes python=${{ matrix.python-version }}
+          mamba install --name fine_env --yes --file requirements_dev.yml
           conda run --name fine_env pip install . --no-deps
           echo "Installation done"
           conda list --name fine_env
