### 2024-04-04 aceeba5
Fix mamba installation
