### 2024-04-04 b036c76
Correct conda info command
