### 2024-04-08 2164e5c
yearlyFullLoadHoursMin/Max can be used to limit annual production
