### 2024-04-08 454a72f
make componentLimit usable with transmission components
