### 2024-04-10 84c9d70
Patch getargspec if not present
