### 2024-04-16 8c03cd0
Add 2dim as parameter and delete doubled method
