+++ fine/transmission.py
@@ -408,16 +408,17 @@ class Transmission(Component):
 
         # operationRateMax
         self.operationRateMax = operationRateMax
+        print(operationRateMax)
         self.fullOperationRateMax = utils.checkAndSetInvestmentPeriodTimeSeries(
-            esM, name, operationRateMax, self.locationalEligibility
+            esM, name, operationRateMax, self.locationalEligibility, "2dim"
         )
         self.aggregatedOperationRateMax = dict.fromkeys(esM.investmentPeriods)
         self.processedOperationRateMax = dict.fromkeys(esM.investmentPeriods)
-
+        print(operationRateFix)
         # operationRateFix
         self.operationRateFix = operationRateFix
         self.fullOperationRateFix = utils.checkAndSetInvestmentPeriodTimeSeries(
-            esM, name, operationRateFix, self.locationalEligibility
+            esM, name, operationRateFix, self.locationalEligibility, "2dim"
         )
         self.aggregatedOperationRateFix = dict.fromkeys(esM.investmentPeriods)
         self.processedOperationRateFix = dict.fromkeys(esM.investmentPeriods)
