### 2024-04-17 2d76e70
add different cases for solver_io
