+++ fine/energySystemModel.py
@@ -2031,7 +2031,10 @@ class EnergySystemModel:
         ################################################################################################################
 
         # Set which solver should solve the specified optimization problem
-        optimizer = opt.SolverFactory(solver, solver_io="python")
+        if solver == "gurobi":
+            optimizer = opt.SolverFactory(solver, solver_io="python")
+        else:
+            optimizer = opt.SolverFactory(solver)
 
         # Set, if specified, the time limit
         if self.solverSpecs["timeLimit"] is not None and solver == "gurobi":
