### 2024-04-18 e0984d5
Check if gurobipy is installed for gurobi python api
