### 2024-04-18 f73f5b2
add test for xrds transmisison input
