### 2024-04-18 9a1fbe4
add test to check xrds to esM conversion
