+++ test/IOManagement/test_xarrayio.py
@@ -254,4 +254,7 @@ def test_transmission_dims(minimal_test_esM):
     xr_dss = xrIO.convertOptimizationInputToDatasets(esM)
     assert (esM.totalTimeSteps == list(xr_dss["Input"]["Transmission"]["Pipelines"].time.to_numpy()))
 
+    esM2 = xrIO.convertDatasetsToEnergySystemModel(xrds)
 
+    capacityMin = esM2.getComponentAttribute("Pipelines", "capacityMin")
+    assert capacityMin.index.name == "space"
