### 2024-04-18 a5e885d
test operationRate for transmission xrds input
