+++ test/IOManagement/test_xarrayio.py
@@ -250,6 +250,18 @@ def test_transmission_dims(minimal_test_esM):
             componentName="Pipelines",
             updateAttrs={"capacityMin": capacityMin},
         )
+    
+    time_index = pd.date_range(start="2020-01-01", periods=4, freq="H")
+    _locs = pd.MultiIndex.from_product([["ElectrolyzerLocation"],["IndustryLocation"]])
+    columns = [f"{idx0}_{idx1}" for idx0, idx1 in _locs]
+    column2 = [f"{idx1}_{idx0}" for idx0, idx1 in _locs]
+    columns = columns + column2
+    operationRateMax = pd.DataFrame(1, index=time_index, columns=columns).reset_index(drop=True)
+    esM.updateComponent(
+            componentName="Pipelines",
+            updateAttrs={"operationRateMax": operationRateMax},
+    )
+    
     esM.optimize()
     xr_dss = xrIO.convertOptimizationInputToDatasets(esM)
     assert (esM.totalTimeSteps == list(xr_dss["Input"]["Transmission"]["Pipelines"].time.to_numpy()))
