### 2024-04-18 7bf23b7
add seperate processing of transmission components in addDFVariablesToXarray
