### 2024-04-18 a4faf13
construct tranmission data in correct format for xarray
