### 2024-04-18 c1b2dfa
adapt addDFVariablesToXarray where transmisison components are now treated sepearately
