+++ test/IOManagement/test_xarrayio.py
@@ -7,6 +7,7 @@ from fine.IOManagement.dictIO import exportToDict
 import fine as fn
 import pandas as pd
 
+
 def compare_values(value_1, value_2):
     """Apply assert functions from pandas if values are pandas.DataFrame or
     pandas.Series, else compare with `==` operator."""
@@ -241,35 +242,41 @@ def test_esm_to_datasets_with_processed_values(minimal_test_esM):
         == 0.177
     )
 
+
 def test_transmission_dims(minimal_test_esM):
-    
     esM = minimal_test_esM
-    capacityMin=pd.DataFrame([[0, 1], [1, 0]], index=list(esM.locations), columns=list(esM.locations))
+    capacityMin = pd.DataFrame(
+        [[0, 1], [1, 0]], index=list(esM.locations), columns=list(esM.locations)
+    )
 
     # update Pipeline component
     esM.updateComponent(
-            componentName="Pipelines",
-            updateAttrs={"capacityMin": capacityMin},
-        )
-    
+        componentName="Pipelines",
+        updateAttrs={"capacityMin": capacityMin},
+    )
+
     time_index = pd.date_range(start="2020-01-01", periods=4, freq="H")
-    _locs = pd.MultiIndex.from_product([["ElectrolyzerLocation"],["IndustryLocation"]])
+    _locs = pd.MultiIndex.from_product([["ElectrolyzerLocation"], ["IndustryLocation"]])
     columns = [f"{idx0}_{idx1}" for idx0, idx1 in _locs]
     column2 = [f"{idx1}_{idx0}" for idx0, idx1 in _locs]
     columns = columns + column2
-    operationRateMax = pd.DataFrame(1, index=time_index, columns=columns).reset_index(drop=True)
-    esM.updateComponent(
-            componentName="Pipelines",
-            updateAttrs={"operationRateMax": operationRateMax},
+    operationRateMax = pd.DataFrame(1, index=time_index, columns=columns).reset_index(
+        drop=True
     )
-    
+    esM.updateComponent(
+        componentName="Pipelines",
+        updateAttrs={"operationRateMax": operationRateMax},
+    )
+
     esM.optimize()
     xr_dss = xrIO.convertOptimizationInputToDatasets(esM)
-    assert (esM.totalTimeSteps == list(xr_dss["Input"]["Transmission"]["Pipelines"].time.to_numpy()))
+    assert esM.totalTimeSteps == list(
+        xr_dss["Input"]["Transmission"]["Pipelines"].time.to_numpy()
+    )
 
     esM2 = xrIO.convertDatasetsToEnergySystemModel(xr_dss)
 
     operationRateMax = esM2.getComponentAttribute("Pipelines", "operationRateMax")
     assert operationRateMax.index.name == "time"
-    
+
     esM2.optimize()
