### 2024-04-18 3fc4259
Modified the existing structure to support a more generalized approach to add new objective functions, added docstrings for additions within component.py
