+++ FINE/component.py
@@ -625,10 +625,9 @@ class Component(metaclass=ABCMeta):
 
         # Set material demand
         self.materialDemandPerCapacity = materialDemandPerCapacity
-        if materialDemandPerCapacity:
-            self.processed_materialDemandPerCapacity = (
-                utils.checkAndSetMaterialDemandPerCapacity(materialDemandPerCapacity)
-            )
+        self.processed_materialDemandPerCapacity = (
+            utils.checkAndSetMaterialDemandPerCapacity(materialDemandPerCapacity)
+        )
 
     def addToEnergySystemModel(self, esM):
         """
@@ -2677,7 +2676,7 @@ class ComponentModel(metaclass=ABCMeta):
         :param objective: type of objective for which the objective function contribution is returned. Default: costs
         :type objective: String
         """
-        # options to select the objective from need to be added here
+        # If new objective functions are implemented, they need to be added here!
         if objective not in ["costs", "material_demand"]:
             raise NotImplementedError("The chosen objective is not supported yet.")
         if objective == "costs":
@@ -2719,73 +2718,108 @@ class ComponentModel(metaclass=ABCMeta):
 
         if objective == "material_demand":
             objective_function_contribution = self.getMaterialDemand(esM, pyM)
-            print("material demand objective got called")
 
         return objective_function_contribution
 
     def getMaterialDemand(self, esM, pyM, getOptValue=False):
-        # TODO Hier attribut mit get_opt_values um optimale Ergebnisse zu bekommen...
-        # TODO dann über commisVar.value -> .value = state der variable, statt variable selbst
-        # selbes design bei getLocEconomicsdesign
-        commisVar = getattr(pyM, "commis_" + self.abbrvName)
+        """
+        Get material demand of a component. This can be used to build the
+        objective function or to get the optimal values after optimizing
+
+        :param esM: EnergySystemModel instance representing the energy system in which the component should be modeled.
+        :type esM: EnergySystemModel instance
+
+        :param pyM: pyomo ConcreteModel which stores the mathematical formulation of the model.
+        :type pyM: pyomo ConcreteModel
+
+        :param getOptValue: Boolean that defines the output of the function:
+
+            - True: Return the optimal material demand values.
+            - False: Return the material demand equation.
+
+            |br| * the default value is False.
+        :type getOptValue: boolean
+
+        :return: material demand equation or optimal material demand values (dictionary containing dataframes for each investment period)
+        :rtype: _type_
+        """
+        component_attribute = "processed_materialDemandPerCapacity"
         if getOptValue:
-            # TODO Add option to retrieve optimal values
-            # TODO Refactor and make it less redundant with the code below
-            material_demand = {}
-            for ip in esM.investmentPeriods:
-                material_demand[ip] = pd.DataFrame()
-            for loc, compName, ip in commisVar:
-                if ip not in esM.investmentPeriods:
-                    continue
-                # check if component has a material demand - if not return NaN
-                # TODO Adapt if this get checked/set better within the corresponding utils function
-                if esM.getComponentAttribute(compName, "materialDemandPerCapacity"):
-                    material_demand[ip].loc[compName, loc] = commisVar[
-                        loc, compName, ip
-                    ].value * esM.getComponentAttribute(
-                        compName, "processed_materialDemandPerCapacity"
-                    )
-                else:
-                    # import pytest
-
-                    # pytest.set_trace()
-                    material_demand[ip].loc[compName, loc] = np.nan
-
-            return material_demand
-        else:
-            # for ip in esM.investmentPeriods:
-            #     for compName in self.componentsDict:
-            #         for loc in esM.locations:
-            #             if self.componentsDict[
-            #                 compName
[…]
