### 2024-04-18 6aca638
subset of data is checked for locations
