+++ .gitlab-ci.yml
@@ -203,6 +203,8 @@ test-codestyle:
   rules:
     # Switch from branch pipeline to merge pipeline once a merge request has
     # been created on the branch.
+    - if: $CI_COMMIT_TAG
+      when: never
     - if: $CI_COMMIT_BRANCH && $CI_OPEN_MERGE_REQUESTS && $CI_PIPELINE_SOURCE == "push"
       when: never
     - if: $CI_COMMIT_BRANCH && $CI_OPEN_MERGE_REQUESTS
@@ -220,7 +222,7 @@ build-tag:
   rules:
     - if: $CI_PIPELINE_SOURCE == "schedule"
       when: never
-    - if: $CI_COMMIT_BRANCH == "master" && $CI_COMMIT_TAG
+    - if: $CI_COMMIT_TAG
 
 # Push develop build to local jugit registry
 build-dev:
@@ -251,5 +253,5 @@ pypi-upload:
     - python3 -m build
     - python3 -m twine upload dist/*
   rules:
-    - if: $CI_COMMIT_BRANCH == "master" && $CI_COMMIT_TAG
+    - if: $CI_COMMIT_TAG
 
