+++ .gitlab-ci.yml
@@ -253,7 +253,7 @@ pypi-upload:
     - test v${PYPROJECT_VERSION} = ${CI_COMMIT_TAG}
     # Build and push to pypi
     - python3 -m build
-    - python3 -m twine upload dist/*
+    - python3 -m twine upload --username $TWINE_USERNAME --password $TWINE_PASSWORD --non-interactive --verbose --disable-progress-bar dist/*
   rules:
     - if: $CI_COMMIT_TAG
 
