+++ .gitlab-ci.yml
@@ -250,7 +250,7 @@ pypi-upload:
     - test v${PYPROJECT_VERSION} = ${CI_COMMIT_TAG}
     # Build and push to pypi
     - python3 -m build
-    - python3 -m twine upload dist/* --username $PYPI_USERNAME --password $PYPI_PASSWORD --non-interactive --verbose --disable-progress-bar
+    - python3 -m twine upload dist/* --username "${PYPI_USERNAME}" --password "${PYPI_PASSWORD}" --non-interactive --verbose --disable-progress-bar
   rules:
     - if: $CI_COMMIT_TAG
 
