+++ .gitlab-ci.yml
@@ -250,7 +250,7 @@ pypi-upload:
     - test v${PYPROJECT_VERSION} = ${CI_COMMIT_TAG}
     # Build and push to pypi
     - python3 -m build
-    - python3 -m twine upload dist/* -u __token__ -p ${TWINE_PASSWORD} --non-interactive --verbose --disable-progress-bar
+    - python3 -m twine upload --repository pypi --non-interactive --verbose --disable-progress-bar dist/*
   rules:
     - if: $CI_COMMIT_TAG
 
