+++ FINE/energySystemModel.py
@@ -22,6 +22,9 @@ warnings.filterwarnings("always", category=UserWarning)
 ####### MOO Additions ######
 from pyaugmecon import PyAugmecon
 import json
+import copy
+import cloudpickle
+from FINE.IOManagement import standardIO
 
 
 class EnergySystemModel:
@@ -2150,60 +2153,64 @@ class EnergySystemModel:
                 warnings.warn("Output is generated for a non-optimal solution.")
             utils.output("\nProcessing optimization output...", self.verbose, 0)
             # Declare component specific sets, variables and constraints
-            w = str(len(max(self.componentModelingDict.keys())) + 6)
+            # w = str(len(max(self.componentModelingDict.keys())) + 6)
 
-            # iterate over investment periods, to get yearly results
-            for key, mdl in self.componentModelingDict.items():
-                if not isinstance(mdl._capacityVariablesOptimum, dict):
-                    mdl._capacityVariablesOptimum = {}
-                __t = time.time()
-                # if _capacityVariablesOptimum is not a dict, convert to dict
-                # (if single year system is optimized several times)
+            # # iterate over investment periods, to get yearly results
+            # for key, mdl in self.componentModelingDict.items():
+            #     if not isinstance(mdl._capacityVariablesOptimum, dict):
+            #         mdl._capacityVariablesOptimum = {}
+            #     __t = time.time()
+            #     # if _capacityVariablesOptimum is not a dict, convert to dict
+            #     # (if single year system is optimized several times)
 
-                mdl.setOptimalValues(self, self.pyM)
-                outputString = (
-                    ("for {:" + w + "}").format(key + " ...")
-                    + "(%.4f" % (time.time() - __t)
-                    + "sec)"
-                )
-                utils.output(outputString, self.verbose, 0)
+            #     mdl.setOptimalValues(self, self.pyM)
+            #     outputString = (
+            #         ("for {:" + w + "}").format(key + " ...")
+            #         + "(%.4f" % (time.time() - __t)
+            #         + "sec)"
+            #     )
+            #     utils.output(outputString, self.verbose, 0)
 
-                # convert optimal values from internal name to external name
-                # e.g. from _capacityVariablesOptimum to capacityVariablesOptimum
-                # For perfectForesight the data stays the same, for a single year optimization
-                # the data is converted from a dict with a single entry to a dataframe
-                # By this, old models will not fail.
-                def convertOptimalValues(esM, mdl, key):
-                    if key in mdl.__dict__.keys():
-                        if esM.numberOfInvestmentPeriods == 1:
-                            setattr(
-                                mdl,
-                                key.replace("_", ""),
-                                getattr(mdl, key)[esM.investmentPeriodNames[0]],
-                            )
-                        else:
-                            setattr(mdl, key.replace("_", ""), getattr(mdl, key))
-                    else:
-                        pass
+            #     # convert optimal values from internal name to external name
+            #     # e.g. from _capacityVariablesOptimum to capacityVariablesOptimum
+            #     # For perfectForesight the data stays the same, for a single year optimization
+            #     # the data is converted from a dict with a single entry to a dataframe
+            #     # By this, old models will not fail.
+            #     def convertOptimalValues(esM, mdl, key):
+            #         if key in mdl.__dict__.keys():
+            #             if esM.numberOfInvestmentPeriods == 1:
+            #                 setattr(
+            #                     mdl,
+            #                     key.replace("_", ""),
+            #                     getattr(mdl, key)[esM.investmentPeriodNames[0]],
+            #                 )
+            #             else:
[…]
