+++ FINE/conversion.py
@@ -383,13 +383,13 @@ class Conversion(Component):
             if self.fullCommodityConversionFactors[ip] != {}:
                 self.aggregatedCommodityConversionFactors[ip] = {}
                 for commod in self.fullCommodityConversionFactors[ip]:
-                    self.aggregatedCommodityConversionFactors[ip][commod] = (
-                        self.getTSAOutput(
-                            self.fullCommodityConversionFactors[ip][commod],
-                            "_commodityConversionFactorTimeSeries" + str(commod) + "_",
-                            data,
-                            ip,
-                        )
+                    self.aggregatedCommodityConversionFactors[ip][
+                        commod
+                    ] = self.getTSAOutput(
+                        self.fullCommodityConversionFactors[ip][commod],
+                        "_commodityConversionFactorTimeSeries" + str(commod) + "_",
+                        data,
+                        ip,
                     )
         else:
             # if depending on the commissioning year, iterate over the relevant commissioning years for the

[…]
