+++ FINE/aggregations/spatialAggregation/__init__.py
@@ -3,6 +3,7 @@ Last edited: February 10, 2022
 
 |br| @author: FINE Developer Team (FZJ IEK-3)
 """
+
 from .aggregation import *
 from .grouping import *
 from .groupingUtils import *
