### 2024-04-24 b4d161b
use _mapC to rebuild transmission connections
