### 2024-04-24 7f8cbfe
locElig is not zero if data is 0. only nan means no data
