### 2024-04-24 1042591
remove declareOperationBinaryVars to prevent gurobi from assuming we have an MILP
