+++ fine/storage.py
@@ -706,8 +706,8 @@ class StorageModel(ComponentModel):
             relevanceThreshold=relevanceThreshold,
         )
         # Operation of component as binary [1/0]
-        self.declareOperationBinaryVars(pyM, "chargeOp_bin")
-        self.declareOperationBinaryVars(pyM, "dischargeOp_bin")
+        # self.declareOperationBinaryVars(pyM, "chargeOp_bin")
+        # self.declareOperationBinaryVars(pyM, "dischargeOp_bin")
         # Capacity development variables [physicalUnit]
         self.declareCommissioningVars(pyM, esM)
         self.declareDecommissioningVars(pyM, esM)
@@ -1540,15 +1540,15 @@ class StorageModel(ComponentModel):
         )
 
         # Operation [physicalUnit*h] is limited by minimum part Load
-        self.additionalMinPartLoad(
-            pyM,
-            esM,
-            "ConstrCharge",
-            "chargeOpConstrSet",
-            "chargeOp",
-            "chargeOp_bin",
-            "cap",
-        )
+        # self.additionalMinPartLoad(
+        #     pyM,
+        #     esM,
+        #     "ConstrCharge",
+        #     "chargeOpConstrSet",
+        #     "chargeOp",
+        #     "chargeOp_bin",
+        #     "cap",
+        # )
 
         #                             Constraints for enforcing discharging operation modes                            #
 
@@ -1583,15 +1583,15 @@ class StorageModel(ComponentModel):
             "processedDischargeOpRateMax",
         )
         # Operation [physicalUnit*h] is limited by minimum part Load
-        self.additionalMinPartLoad(
-            pyM,
-            esM,
-            "ConstrDischarge",
-            "dischargeOpConstrSet",
-            "dischargeOp",
-            "dischargeOp_bin",
-            "cap",
-        )
+        # self.additionalMinPartLoad(
+        #     pyM,
+        #     esM,
+        #     "ConstrDischarge",
+        #     "dischargeOpConstrSet",
+        #     "dischargeOp",
+        #     "dischargeOp_bin",
+        #     "cap",
+        # )
 
         # Cyclic constraint enforcing that all storages have the same state of charge at the the beginning of the first
         # and the end of the last time step
