+++ FINE/energySystemModel.py
@@ -2152,65 +2152,9 @@ class EnergySystemModel:
             ):
                 warnings.warn("Output is generated for a non-optimal solution.")
             utils.output("\nProcessing optimization output...", self.verbose, 0)
-            # Declare component specific sets, variables and constraints
-            # w = str(len(max(self.componentModelingDict.keys())) + 6)
 
-            # # iterate over investment periods, to get yearly results
-            # for key, mdl in self.componentModelingDict.items():
-            #     if not isinstance(mdl._capacityVariablesOptimum, dict):
-            #         mdl._capacityVariablesOptimum = {}
-            #     __t = time.time()
-            #     # if _capacityVariablesOptimum is not a dict, convert to dict
-            #     # (if single year system is optimized several times)
-
-            #     mdl.setOptimalValues(self, self.pyM)
-            #     outputString = (
-            #         ("for {:" + w + "}").format(key + " ...")
-            #         + "(%.4f" % (time.time() - __t)
-            #         + "sec)"
-            #     )
-            #     utils.output(outputString, self.verbose, 0)
-
-            #     # convert optimal values from internal name to external name
-            #     # e.g. from _capacityVariablesOptimum to capacityVariablesOptimum
-            #     # For perfectForesight the data stays the same, for a single year optimization
-            #     # the data is converted from a dict with a single entry to a dataframe
-            #     # By this, old models will not fail.
-            #     def convertOptimalValues(esM, mdl, key):
-            #         if key in mdl.__dict__.keys():
-            #             if esM.numberOfInvestmentPeriods == 1:
-            #                 setattr(
-            #                     mdl,
-            #                     key.replace("_", ""),
-            #                     getattr(mdl, key)[esM.investmentPeriodNames[0]],
-            #                 )
-            #             else:
-            #                 setattr(mdl, key.replace("_", ""), getattr(mdl, key))
-            #         else:
-            #             pass
-
-            #     optimalValueParameters = [
-            #         "_optSummary",
-            #         "_stateOfChargeOperationVSariablesOptimum",
-            #         "_chargeOperationVariablesOptimum",
-            #         "_dischargeOperationVariablesOptimum",
-            #         "_phaseAngleVariablesOptimum",
-            #         "_operationVariablesOptimum",
-            #         "_discretizationPointVariablesOptimun",
-            #         "_discretizationSegmentConVariablesOptimun",
-            #         "_discretizationSegmentBinVariablesOptimun",
-            #         "_capacityVariablesOptimum",
-            #         "_isBuiltVariablesOptimum",
-            #         "_commissioningVariablesOptimum",
-            #         "_decommissioningVariablesOptimum",
-            #     ]
-
-            #     for optParam in optimalValueParameters:
-            #         convertOptimalValues(self, mdl, optParam)
-
-            self.process_fine_output(
-                solved_pyomo_model=self.pyM
-            )  # TODO STILL NEEDS TO BE TESTED!
+            # Process the pyomo-based FINE optimization output
+            self.process_fine_output(solved_pyomo_model=self.pyM)
 
             # Store the objective value in the EnergySystemModel instance.
             self.objectiveValue = self.pyM.Obj()
@@ -2384,10 +2328,12 @@ class EnergySystemModel:
     ):
         # TODO Add docstrings
 
-        # Normally this gets set within esM.optimize() therefore we have to set it manually here otherwise it will be missed
-        # TODO Find out if theres a way to make everything work with TSA otherwise its going to be a calcualation nightmare...
+        # TODO This function does currently only work with a customized fork of the original PyAugmecon package
[…]
