### 2024-06-03 f7fda88
bug fix for multi period data for transmission components
