+++ test/perfectForesight/test_perfectForesight_miniTest.py
@@ -213,6 +213,14 @@ def test_perfectForesight_storage_transmission(perfectForesight_test_esM):
             commodity="hydrogen",
             hasCapacityVariable=True,
             investPerCapacity=0.177,
+            capacityMax={
+                ip: pd.DataFrame(
+                    data=[[0.2, 0.2], [0.25, 0.25]],
+                    index=list(esM.locations),
+                    columns=list(esM.locations)
+                ) * (1+(ip-2020)/40)
+                for ip in esM.investmentPeriodNames
+            },
             interestRate=0.08,
             economicLifetime=40,
         )
