+++ fine/energySystemModel.py
@@ -25,7 +25,8 @@ from pyaugmecon import PyAugmecon
 import json
 import copy
 import cloudpickle
-from FINE.IOManagement import standardIO
+from fine.IOManagement import standardIO
+from typing import Union, List
 
 
 class EnergySystemModel:
@@ -91,7 +92,9 @@ class EnergySystemModel:
         balanceLimit=None,
         pathwayBalanceLimit=None,
         annuityPerpetuity=False,
-        objective="costs",  # maybe not the best option...TODO Think about this flag
+        objective: Union[
+            str, List[str]
+        ] = "costs",  # maybe not the best option...TODO Think about this flag
     ):
         """
         Constructor for creating an EnergySystemModel class instance
@@ -2061,7 +2064,7 @@ class EnergySystemModel:
         ################################################################################################################
 
         # Set which solver should solve the specified optimization problem
-        if solver == "gurobi" and importlib.util.find_spec('gurobipy'):
+        if solver == "gurobi" and importlib.util.find_spec("gurobipy"):
             # Use the direct gurobi solver that uses the Python API.
             optimizer = opt.SolverFactory(solver, solver_io="python")
         else:
@@ -2282,9 +2285,6 @@ class EnergySystemModel:
             # (if single year system is optimized several times)
 
             mdl.setOptimalValues(self, solved_pyomo_model)
-            # import pytest
-
-            # pytest.set_trace()
             outputString = (
                 ("for {:" + w + "}").format(key + " ...")
                 + "(%.4f" % (time.time() - __t)
@@ -2334,6 +2334,8 @@ class EnergySystemModel:
         pyaugmeconOptions,
         declaresOptimizationProblem=True,
         deletePyomoModels=True,
+        write_excel_output=True,
+        excel_output_path=None,
     ):
         # TODO Add docstrings
 
@@ -2372,9 +2374,90 @@ class EnergySystemModel:
         # Deactivate all objectives by default - this is PyAugmecon-specific!
         for o in range(len(self.pyM.obj_list)):
             self.pyM.obj_list[o + 1].deactivate()
+
+        print("########## CALLING PY AUGMECON #######")
         pyaugmecon = PyAugmecon(self.pyM, pyaugmeconOptions)
         pyaugmecon.solve()  # solve PyAugmecon multi-objective optimization problem
 
+        # def load_pyomo_model_from_file(file_name, custom_path=None, delete_file=False):
+        #     """Load a "pickled" pyomo model instance from a file.
+        #     Important: To use the imported model and to process its output with FINE
+        #     it has to be compatible with the current FINE-model instance.
+
+        #     Args:
+        #         file_name (str): name of file that contains pickled pyomo model instance
+
+        #     Returns:
+        #         pyomo.ConcreteModel: The unpickled pyomo model instance.
+        #     """
+
+        #     file_path = file_name
+        #     if custom_path:
+        #         if not os.path.isdir(custom_path):
+        #             os.mkdir(custom_path)
+        #         file_path = os.path.join(custom_path, file_name)
+        #     with open(f"{file_path}.pkl", mode="rb") as file:
+        #         pyomo_model = cloudpickle.load(file)
+        #     if delete_file:
+        #         os.remove(f"{file_path}.pkl")
+        #     return pyomo_model
+
+        # # delete existing pyomo model from esm to make the esm "copyable"
+        # del self.pyM
+
+        # # for each solution (i.e. pareto point) copy the existing esm instance,
+        # # load the corresponding pyomo model instance and assign it to the pyM
+        # # instance of the copied esm model, use the default FINE result post-processing
+        # # and delete the copied esm instance when done.
[…]
