### 2024-06-17 685fc32
add Conversion (rudimentary) to componentLimit
