### 2024-06-17 f0888bf
add performance summary to netcdf files
