### 2024-07-01 7a82189
added possibility to pass custom gurobi optimization specs to pyaugmecon
