+++ fine/energySystemModel.py
@@ -1731,6 +1731,7 @@ class EnergySystemModel:
             utils.output(
                 "Initializing multi-objective optimization...", self.verbose, 0
             )
+            # Add create an objective list and add all objectives to it
             pyM.obj_list = pyomo.ObjectiveList()
             for (
                 obj
@@ -2333,37 +2334,70 @@ class EnergySystemModel:
         self,
         pyaugmeconOptions,
         declaresOptimizationProblem=True,
+        timeSeriesAggregation=False,
+        solver="gurobi",
+        optimizationSpecs: Union[dict, str] = None,
         deletePyomoModels=True,
         write_excel_output=True,
         excel_output_path=None,
     ):
         # TODO Add docstrings
+        # Future TODO add a performance summary similar to the one within esm.optimize()
+        # Future TODO maybe use the utils.checkOptimizeInput() function to also check the input here or implement a similar function
 
-        # TODO This function does currently only work with a customized fork of the original PyAugmecon package
+        # TODO This method does currently only work with a customized fork of the original PyAugmecon package
         # This fork can be found here: https://github.com/kschulze26/pyaugmecon/tree/add_model_export
 
-        self.segmentation = False  # Normally this gets set within esM.optimize() therefore we have to set it manually here otherwise it will be missed
-        # TODO Find out if there is a way to make everything work with TSA otherwise its going to be a calcualation nightmare...
-        # TODO implement the switch based on the argument timeSeriesAggregation as for example used within the "declareOptimizationProblem"
+        # Set the segmentation parameter to false if time series aggregation is disabled
+        if not timeSeriesAggregation:
+            self.segmentation = False
 
+        # Declare optimization problem if not already done
         if declaresOptimizationProblem:
-            self.declareOptimizationProblem()
-            # TODO pass more options/arguments to declareOptimizationProblem
+            self.declareOptimizationProblem(timeSeriesAggregation)
         else:
+            # Throw an error if declareOptimizationProblem is False but no pyomo model exists (same error as in optimize())
             if self.pyM is None:
                 raise TypeError(
                     "The optimization problem is not declared yet. Set the argument declaresOptimization"
                     " problem to True or call the declareOptimizationProblem function first."
                 )
 
-        ################ PYAUGMECON ################
+        ############################################################################
+        ################################ PYAUGMECON ################################
+        ############################################################################
+
         # This section deals with the PyAugmecon-specific operations
         # For a detailed documentation of the corresponding package please
         # refer to https://github.com/wouterbles/pyaugmecon
 
+        # Integrate fine solver options and optimization specs into pyaugmecon options
+
+        # Set which solver should solve the specified optimization problem
+        if solver == "gurobi" and importlib.util.find_spec("gurobipy"):
+            # Use the direct gurobi solver that uses the Python API.
+            pyaugmeconOptions["solver_io"] = "python"
+        else:
+            pyaugmeconOptions["solver_io"] = None
+
+        if not pyaugmeconOptions.get("solver_name"):
+            pyaugmeconOptions["solver_name"] = solver
+        else:
+            if pyaugmeconOptions["solver_name"] != solver:
+                raise ValueError(
+                    f'Two different solver name were passed! \n solver defined in pyaugmecon options: {pyaugmeconOptions["solver_name"]} \n solver defined in esm.optimize_moo(): {solver}.'
+                )
+
+        # Convert optimization specs from string to dictionary
+        if optimizationSpecs:
[…]
