### 2024-07-15 fd8aca5
add clustered input time series from TSA to xarray to enable calculation of curtailment
