+++ paper/paper.md
@@ -12,6 +12,12 @@ authors:
   - name: Kevin Knosala
     orcid: 0000-0002-8825-5934
     affiliation: "1, 2"
+  - name: Johannes Behrens
+    orcid: 0009-0003-2396-0282
+    affiliation: "1, 2"
+  - name: Rachel Maier
+    orcid: 0000-0003-0664-3886
+    affiliation: "1, 2"
   - name: Maximilian Hoffmann
     orcid: 0000-0003-1232-8110
     affiliation: "1"
@@ -22,53 +28,58 @@ authors:
     orcid: 0000-0002-1671-3262
     affiliation: "1, 2"
 affiliations:
-  - name: Forschungszentrum Jülich GmbH, Institute of Energy and Climate Research – Techno-economic Systems Analysis (IEK-3), 52425 Jülich, Germany
+  - name: Forschungszentrum Jülich GmbH, Institute of Climate and Energy Systems – Jülich Systems Analysis, 52425 Jülich, Germany
     index: 1
   - name: RWTH Aachen University, Chair for Fuel Cells, Faculty of Mechanical Engineering, 52062 Aachen, Germany
     index: 2
-date: 13 October 2023
+date: 01 July 2024
 bibliography: FINE-literature.bib
 ---
 
 # Summary
 
-The decarbonization of energy systems worldwide requires a transformation of their design and operation across all sectors, that is, the residential and commercial, industrial, and transportation sectors. Energy system models are frequently employed for assessing these changes, providing scenarios on potential future system design and on how new technologies and a modified infrastructure will meet future energy demands. Thus, they support investment decisions and policy-making.
-The Python-based Framework for Integrated Energy System Assessment (`ETHOS.FINE`) is a software package that provides a toolbox for modeling, analyting and evaluating such energy systems using mathematical optimization. `ETHOS.FINE` is part of the Energy Transformation paTHway Optimization Suite (`ETHOS`)[^1], a collection of modeling tools developed by the Institute of Energy and Climate Research - Techno-Economic System Analysis (IEK-3) at Forschungszentrum Jülich. `ETHOS` offers a holistic view on energy systems at arbitrary scales providing tools for geospatial renewable potential analyses, time series simulation tools for residential and industrial sector, discrete choice models for the transportation sector, modeling of global energy supply routes, and local infrastructure assessments, among others. The `ETHOS` model suite is, e.g., used for analyzing the energy transition of Germany [@Stolten2022].
+The decarbonization of energy systems worldwide requires a transformation of their design and operation across all sectors, that is, the residential and commercial, industrial, and transportation sectors. Energy system models are frequently employed for assessing these changes, providing scenarios on potential future system design and on how new technologies and a modified infrastructure will meet future energy demands. Thus, they support investment decisions and policy-making. The Python-based Framework for Integrated Energy System Assessment (`ETHOS.FINE`) is a software package that provides a toolbox for modeling, analyzing and evaluating such energy systems using mathematical optimization. 
 
-`ETHOS.FINE` is not limited to a single instance of energy system. Instead, it can be freely adapted to consider multiple commodities, regions, time steps and investment periods. The optimization objective is to minimize the total annual cost of the system and is subject to technical and environmental constraints. The generic object-oriented implementation allows for arbitrary spatial scales and numbers of regions – from the local level, e.g., individual buildings, to the regional one, e.g., districts or industrial sites, to the national and international levels. Furthermore, the spatial technological resolution can be aggregated by built-in aggregation methods that are described in @Patil_2022 for reducing model size and complexity.
[…]
