+++ paper/FINE-literature.bib
@@ -80,6 +80,8 @@ publisher={Springer},
 DOI = {10.1007/s12532-011-0026-8}
 }
 
+@book{bynum2021pyomo, title={Pyomo--optimization modeling in python}, author={Bynum, Michael L. and Hackebeil, Gabriel A. and Hart, William E. and Laird, Carl D. and Nicholson, Bethany L. and Siirola, John D. and Watson, Jean-Paul and Woodruff, David L.}, edition={Third}, volume={67}, year={2021}, publisher={Springer Science \& Business Media} }
+
 @article{Hoffmann_2020,
  author = {Hoffmann, Maximilian and Kotzur, Leander and Stolten, Detlef and Robinius, Martin},
  doi = {10.3390/en13030641},
@@ -285,16 +287,43 @@ Type: https://doi.org/10.1016/j.rset.2023.100059},
 
 @phdthesis{Welder2022,
       author       = {Welder, Lara},
-      title        = {{O}ptimizing {C}ross-linked {I}nfrastructure for {F}uture {E}nergy {S}ystems},
+      title        = {{O}ptimizing {C}ross-linked {I}nfrastructure for {F}uture
+                      {E}nergy {S}ystems},
       volume       = {591},
       school       = {RWTH Aachen University},
       type         = {Dissertation},
       address      = {Jülich},
-      publisher    = {Forschungszentrum Jülich GmbH Zentralbibliothek, Verlag},
-      isbn         = {978-3-95806-661-8},
-      series       = {Schriften des Forschungszentrums Jülich. Reihe Energie
-                      $\&$ Umwelt / Energy $\&$ Environment},
+      publisher    = {Forschungszentrum Jülich GmbH Zentralbibliothek Verlag},
+      reportid     = {FZJ-2022-06354},
+      isbn         = {978-3-95806-659-5},
+      series       = {Schriften des Forschungszentrums Jülich Reihe Energie $\&$
+                      Umwelt / Energy $\&$ Environment},
+      pages        = {xxiii, 360},
       year         = {2022},
+      note         = {Dissertation, RWTH Aachen University, 2022},
+      abstract     = {Increasing shares of distributed, intermittent renewable
+                      electricity generation as well as the need to decarbonize
+                      all sectors in future energy systems require new
+                      infrastructure solutions. In these infrastructure solutions,
+                      energy transmission and storage technologies will serve as
+                      spatial and temporal balancing options. Technologies like
+                      electrolyzers will be deployed to create a renewable link
+                      between the power sector and other sectors. Together with
+                      technologies for power generation from renewably produced
+                      gases, electricity and gas infrastructurewill be strongly
+                      cross-linked. The design of such cross-linked infrastructure
+                      is a computationally challenging undertaking and national
+                      cross-linked infrastructure analyses are scarce in
+                      literature....},
+      cin          = {IEK-3},
+      cid          = {I:(DE-Juel1)IEK-3-20101013},
+      pnm          = {1111 - Effective System Transformation Pathways (POF4-111)
+                      / 1112 - Societally Feasible Transformation Pathways
+                      (POF4-111)},
+      pid          = {G:(DE-HGF)POF4-1111 / G:(DE-HGF)POF4-1112},
+      typ          = {PUB:(DE-HGF)3 / PUB:(DE-HGF)11},
[…]
