### 2024-07-31 b7e0f4d
adds support for ip-based moo attributes (e.g. specific material requirements that are not static) - still needs more testing
