### 2024-08-01 FZJ-IEK3-VSA/FINE#47 @t-gross
Hi, I apologize that it took so long to work on the paper issues.
We have put the installation instructions on our readTheDocs page. So, it is also possible to access it via the menu.
In the ReadMe, there is only a short description with the recommended way to install and a link to the more detailed description on readTheDocs.

Thanks for raising the issue.
Kind regards,
@t-gross
