### 2024-08-06 c8c8c0a
342 improve performance xarray loading
