### 2024-08-06 b69903d
Introduce mandatory linting for FINE
