### 2024-08-06 37ff838
Introduce mandatory linting for FINE
