+++ paper/FINE-literature.bib
@@ -634,7 +634,7 @@ Type: https://doi.org/10.1016/j.rset.2023.100059},
 }
 
 @article{oemof,
-	doi = {https://doi.org/10.1016/j.esr.2018.07.001},
+	doi = {10.1016/j.esr.2018.07.001},
 	year = {2018},
 	journal = {Energy Strategy Reviews},
 	author = {S. Hilpert and C. Kaldemeyer and U. Krien and S. G\"unther and C. Wingenbach and G. Plessmann},
