### 2024-08-26 #6274 @t-gross
> 👋 @t-gross – it's now the end of August. How are you getting on?

Thanks for the question, and again my apologies that it took so long.
I just uploaded the latest version of the paper (https://github.com/FZJ-IEK3-VSA/FINE/tree/paper).
We tried to address most of the comments of our reviewers. I also wanted to thank them to read and comment on our paper draft and our software package.

As an update on this:

Repository:
- We published a new version of our package with some new major features (adding investment periods as additional time dimension). This enables the modeler to optimize the whole transition pathway of an energy system.
- We reorganized the examples and added the description to the readme.
- We added a new section in the readme about how to contribute and get support. Additionally, we added a discussion area and the code of conduct.
- Installation description was updated; mamba is still the recommended way to install the package.
- Added specifiation of versions for package's dependencies

Paper:
- Update of the paper regarding the new features.
-  Due to this new features,  two additional authors have been added who are primarily responsible for the development of these features.
- Paragraph about ETHOS model suite has been moved to another section (summary -> examples of previous usage).
- Paragraph about built-in aggregation methods has been revised.
- Examples of other framework tools have been added.
- Added sentence about the usage of pyomo (historical reason; might be changed in future to improve performance).
- Update references (correct links)

Not added:
- As the performance of the model setup and optimization process is highly dependent on the computational resources, we have not added performance indicators to the publication.

Ongoing:
- Describing the mathematical background of the model is currently in progress and will be published on ReadTheDocs (https://vsa-fine.readthedocs.io/en/latest/index.html) (connected to https://github.com/FZJ-IEK3-VSA/FINE/issues/51). As the software package is further developed, new boundary conditions, parameters and variables must be added. We are working on this issue and hope to be able to put a first version online soon.
