### 2024-09-05 023f7d1
Fix decommissioning constraint
