### 2024-09-05 a7ae97d
359 gurobi issue
