### 2024-09-05 f39b51c
359 gurobi issue
