+++ fine/conversion.py
@@ -1000,7 +1000,7 @@ class ConversionModel(ComponentModel):
             isOperationCommisYearDepending=True,
         )
 
-        # Operation [physicalUnit*h] is limited by minimum part Load
+        # # Operation [physicalUnit*h] is limited by minimum part Load
         self.additionalMinPartLoad(
             pyM, esM, "ConstrOperation", "opConstrSet", "op", "op_bin", "cap"
         )
