### 2024-09-27 96c4b80
Allow pathway parameter for single investment periods
