+++ fine/utils.py
@@ -565,16 +565,6 @@ def processBoundParams(esM, param):
     return processedParam
 
 
-def checkForConsistency(paramName, valueName, name, capacityBound):
-    if isinstance(capacityBound, dict):
-        if any(x is not None for x in capacityBound.values()):
-            if not all(x is not None for x in capacityBound.values()):
-                raise ValueError(
-                    "A mix between None and specified values is not allowed "
-                    + f"between investment periods for {paramName}{valueName} for {name}."
-                )
-
-
 def checkAndSetBounds(esM, name, paramName, MinVal, MaxVal, FixVal):
     checkInvestmentPeriodParameters(name, MinVal, esM.investmentPeriodNames)
     checkInvestmentPeriodParameters(name, MaxVal, esM.investmentPeriodNames)
@@ -626,11 +616,6 @@ def checkAndSetBounds(esM, name, paramName, MinVal, MaxVal, FixVal):
                     f"{paramName}Fix values < {paramName}Min values detected."
                 )
 
-    # check if there is a mix of None and specified boundaries in one of the capacityBounds
-    checkForConsistency(paramName, "Max", name, MaxVal)
-    checkForConsistency(paramName, "Min", name, MinVal)
-    checkForConsistency(paramName, "Fix", name, FixVal)
-
     return processedMinVal, processedMaxVal, processedFixVal
 
 
@@ -645,12 +630,6 @@ def checkInvestmentPeriodParameters(name, param, years):
                 f"'{name}' has different ip-names ('{param.keys()}')"
                 + f" than the investment periods of the esM ('{years}')",
             )
-
-        for key, value in param.items():
-            if value is None:
-                raise ValueError(
-                    f"Currently a dict containing None values cannot be passed for '{name}'"
-                )
 
 
 def checkCapacityDevelopmentWithStock(
@@ -699,14 +678,14 @@ def checkCapacityDevelopmentWithStock(
         # stock capacity
         for loc in stockCapacity.columns:
             for year in investmentPeriods:
-                if capacityMax is not None:
+                if capacityMax[year] is not None:
                     if stockCapacity.loc[year, loc] > capacityMax[year][loc]:
                         raise ValueError(
                             "Mismatch between stock capacity (by its "
                             + "commissioning and the technical lifetime) and "
                             + "capacityMax"
                         )
-                if capacityFix is not None:
+                if capacityFix[year] is not None:
                     if stockCapacity.loc[year, loc] > capacityFix[year][loc]:
                         raise ValueError(
                             "Mismatch between stock capacity (by its "
@@ -937,17 +916,10 @@ def setLocationalEligibility(
         def defineLocDependencyCapacityBounds(name, capacityBound):
             if capacityBound is None:
                 return False
-            anyLocIndependent = any(
-                x is None or isinstance(x, (int, float)) for x in capacityBound.values()
-            )
             anyLocDependent = any(
                 x is not None and not isinstance(x, (int, float))
                 for x in capacityBound.values()
             )
-            if anyLocDependent and anyLocIndependent:
-                raise ValueError(
-                    f"Please implement {name} either as location dependent or indendent consistent over entire pathway."
-                )
             if anyLocDependent:
                 return True
             else:
@@ -963,12 +935,17 @@ def setLocationalEligibility(
         if isinstance(operationTimeSeries, dict) and len(operationTimeSeries) == 0:
             operationTimeSeries = None
 
-        if not hasCapacityVariable and operationTimeSeries is not None:
+        if (
+                not hasCapacityVariable
+                and operationTimeSeries is not None
+                and any(ots is not None for ots in operationTimeSeries.values())
[…]
