### 2024-09-27 d5e77cb
Allow pathway parameter for single investment periods
