### 2024-10-09 5fb732c
Bug fix in quadratic formulation
