### 2024-10-16 d01fd00
removes pytest.set_trace() statement
