### 2024-10-23 a515524
Resolve "Package ipykernel missing when installing fine from conda-forge"
