+++ requirements_dev.yml
@@ -24,6 +24,7 @@ dependencies:
   - pwlf<3
   - psutil<6
   - gurobi-logtools<4
+  - ipykernel<=6.29.5
   # develop
   - sphinx<8
   - sphinx_rtd_theme<3
@@ -33,3 +34,4 @@ dependencies:
   - pytest-xdist<4
   - nbval<1
   - ruff<1
+  
