### 2024-10-23 10e6851
Resolve "Package ipykernel missing when installing fine from conda-forge"
