### 2024-10-24 fa9aa6d
Add option for flexible conversion components
