+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "fine"
-version = "2.3.8"
+version = "2.3.9"
 description = "Framework for integrated energy systems assessment"
 readme = "README.md"
 authors = [
@@ -30,22 +30,22 @@ classifiers = [
 ]
 keywords = ["energy assesment", "energy system", "optimization"]
 dependencies = [
-    "geopandas<1",
-    "openpyxl<4",
-    "matplotlib<4",
-    "xlrd<3",
-    "pyomo<7",
-    "numpy<2",
-    "pandas>=2,<3",
-    "scipy<2",
-    "scikit-learn>=1.2,<2",
+    "geopandas<=0.14.4",
+    "openpyxl<=3.1.5",
+    "matplotlib<=3.9.2",
+    "xlrd<=2.0.1",
+    "pyomo<=6.8.0",
+    "numpy<=1.26.4",
+    "pandas>=2,<=2.2.3",
+    "scipy<=1.14.1",
+    "scikit-learn>=1.2,<=1.5.2",
     "xarray<=2024.3",
-    "rasterio<2",
-    "netcdf4<2",
+    "rasterio<=1.4.1",
+    "netcdf4<=1.7.2",
     "tsam",
-    "pwlf<3",
-    "psutil<6",
-    "gurobi-logtools<4",
+    "pwlf<=2.2.1",
+    "psutil<=5.9.8",
+    "gurobi-logtools<=3.1.0",
     "ipykernel<=6.29.5",
 ]
 
@@ -53,14 +53,14 @@ requires-python = ">=3.10,<3.13"
 
 [project.optional-dependencies]
 develop = [
-    "sphinx<8",
-    "sphinx_rtd_theme<3",
-    "myst-parser<3",
-    "pytest<9",
-    "pytest-cov<5",
-    "pytest-xdist<4",
-    "nbval<1",
-    "ruff<1",
+    "sphinx<=7.4.4",
+    "sphinx_rtd_theme<=2.0.0",
+    "myst-parser<=2.0.0",
+    "pytest<=8.3.3",
+    "pytest-cov<=4.1.0",
+    "pytest-xdist<=3.6.1",
+    "nbval<=0.11.0",
+    "ruff<0.7.0",
 ]
 
 #Configureation options
