### 2024-11-14 a8d7f34
adds small change to type check for material demands
