### 2024-11-14 06a2d66
make partloadMin work with segmentation
