### 2024-11-18 3e91552
fix usage of hoursPerTimeStep for capacity
the maximum capacity and chargeRate should not be depending on the hoursPerTimeStep as it is a total constraint
