### 2024-11-18 5379108
fixes problems for single year optimizations, improves test for ip dependent material demands
