+++ fine/energySystemModel.py
@@ -257,6 +257,9 @@ class EnergySystemModel:
             |br| * the default value is False
         :type: annuityPerpetuity: bool
 
+        :param objective: the objective(s) of the energy system model. Currently supported are "costs" and "material_demand".
+        :type: objective: str or List[str]
+
         """
 
         # Check correctness of inputs
@@ -424,9 +427,17 @@ class EnergySystemModel:
         }
         self.objectiveValue = None
 
-        #### MultiObjective Additions ####
-        self.objective = objective  # TODO Add check if its a string or a list
-        if type(self.objective) == list:
+        ################################################################################################################
+        #                                         Multi-objective additions                                            #
+        ################################################################################################################
+        # The objective parameter defines the objectives of the energy system model
+        # If only one objective is passed the workflow for single objective optimization is used and the parameter
+        # use_moo is set to False.
+        # If a list of objectives is passed, e.g. ["costs", "material_demand"] the workflow for multi-objective
+        # optimization is used and the parameter use_moo is set to True
+
+        self.objective = objective
+        if isinstance(self.objective, list):
             self.use_moo = True
         else:
             self.use_moo = False
@@ -1241,17 +1252,14 @@ class EnergySystemModel:
 
             def initIntraYearTimeSet(pyM):
                 return (
-                    (p, t)
-                    for p in self.typicalPeriods
-                    for t in self.timeStepsPerPeriod
+                    (p, t) for p in self.typicalPeriods for t in self.timeStepsPerPeriod
                 )
 
             def initInvestPeriodInterPeriodSet(pyM):
                 return (
                     (t_inter)
                     for t_inter in range(
-                        int(len(self.totalTimeSteps) / len(self.timeStepsPerPeriod))
-                        + 1
+                        int(len(self.totalTimeSteps) / len(self.timeStepsPerPeriod)) + 1
                     )
                 )
 
@@ -1287,17 +1295,14 @@ class EnergySystemModel:
 
             def initIntraYearTimeSet(pyM):
                 return (
-                    (p, t)
-                    for p in self.typicalPeriods
-                    for t in self.segmentsPerPeriod
+                    (p, t) for p in self.typicalPeriods for t in self.segmentsPerPeriod
                 )
 
             def initInvestPeriodInterPeriodSet(pyM):
                 return (
                     (t_inter)
                     for t_inter in range(
-                        int(len(self.totalTimeSteps) / len(self.timeStepsPerPeriod))
-                        + 1
+                        int(len(self.totalTimeSteps) / len(self.timeStepsPerPeriod)) + 1
                     )
                 )
 
@@ -1353,7 +1358,9 @@ class EnergySystemModel:
             # iterate over balance limit to define either minimal, maximal or both balance limits per balanceLimitID
             for ip in self.investmentPeriods:
                 if self.processedBalanceLimit[ip] is not None:
-                    for balanceLimitID, data in self.processedBalanceLimit[ip].iterrows():
+                    for balanceLimitID, data in self.processedBalanceLimit[
+                        ip
+                    ].iterrows():
                         # check for regional constraints
                         for loc in self.locations:
                             if data[loc] is not None:
@@ -2333,15 +2340,11 @@ class EnergySystemModel:
         timeSeriesAggregation=False,
         solver="gurobi",
         optimizationSpecs: Union[dict, str] = None,
-        deletePyomoModels=True,
-        write_excel_output=True,
-        excel_output_path=None,
     ):
         # TODO Add docstrings
         # Future TODO add a performance summary similar to the one within esm.optimize()
-        # Future TODO maybe use the utils.checkOptimizeInput() function to also check the input here or implement a similar function
 
-        # TODO This method does currently only work with a customized fork of the original PyAugmecon package
[…]
