+++ fine/expansionModules/transformationPath.py
@@ -158,9 +158,7 @@ def optimizeSimpleMyopic(
 
         if trackESMs:
             tmp = esM
-            del (
-                tmp.pyM
-            )  # Delete pyomo instance from esM before copying it (pyomo instances cannot be copied)
+            del tmp.pyM  # Delete pyomo instance from esM before copying it (pyomo instances cannot be copied)
             myopicResults.update({"ESM_" + str(mileStoneYear): copy.deepcopy(tmp)})
 
         # Get stock if not all optimizations are done
