### 2024-11-19 8199f50
Bug fix for optimization summary of storage components
