+++ fine/IOManagement/xarrayIO.py
@@ -107,55 +107,110 @@ def convertOptimizationOutputToDatasets(esM, optSumOutputLevel=0):
             optSum = esM.getOptimizationSummary(name, ip=ip, outputLevel=oL_)
             if esM.componentModelingDict[name].dimension == "1dim":
                 for component in optSum.index.get_level_values(0).unique():
-                    variables = optSum.loc[component].index.get_level_values(0)
-                    units = optSum.loc[component].index.get_level_values(1)
-                    variables_unit = dict(zip(variables, units))
-
                     for variable in (
                         optSum.loc[component].index.get_level_values(0).unique()
-                    ):
-                        df = optSum.loc[(component, variable)]
-                        df = df.iloc[-1]
-                        df.name = variable
-                        df.index.rename("space", inplace=True)
-                        df = pd.to_numeric(df)
-                        xr_da = df.to_xarray()
-
-                        # add variable [e.g. 'TAC'] and units to attributes of xarray
-                        unit = variables_unit[variable]
-                        xr_da.attrs[variable] = unit
-
+                        ):
+                        df_o = optSum.loc[(component, variable)]
+                        # differentiate if two entries per variable
+                        if df_o.shape[0]==2:
+                            df = df_o.iloc[0].copy()
+                            df.name = variable
+                            df.index.rename("space", inplace=True)
+                            df = pd.to_numeric(df)
+                            xr_da = df.to_xarray()
+                            unit = df_o.iloc[0].name
+                            # unit = variables_unit[variable]
+                            xr_da.attrs[variable] = unit
+                            # merge to overall xr_dss
+                            xr_dss[ip][name][component] = xr.merge(
+                            [xr_dss[ip][name][component], xr_da],
+                            combine_attrs="drop_conflicts",
+                            )                           
+                            # if a variable occurs twice keep both in separate lines, example: operation_annual, operation
+                            df = df_o.iloc[1].copy()
+                            df.name = f"{variable}_{1}"
+                            df.index.rename("space", inplace=True)
+                            df = pd.to_numeric(df)
+                            xr_da = df.to_xarray()
+                            # add variable [e.g. 'TAC'] and units to attributes of xarray
+                            unit = df_o.iloc[1].name
+                            xr_da.attrs[df.name] = unit    
+                        else:
+                            df = df_o.iloc[-1]
+                            df.name = variable
+                            df.index.rename("space", inplace=True)
+                            df = pd.to_numeric(df)
+                            xr_da = df.to_xarray()
+                            # add variable [e.g. 'TAC'] and units to attributes of xarray
+                            unit = df_o.iloc[-1].name
+                            xr_da.attrs[variable] = unit
+                        
+                        # merge to overall xr_ds
                         xr_dss[ip][name][component] = xr.merge(
                             [xr_dss[ip][name][component], xr_da],
                             combine_attrs="drop_conflicts",
                         )
             elif esM.componentModelingDict[name].dimension == "2dim":
                 for component in optSum.index.get_level_values(0).unique():
-                    variables = optSum.loc[component].index.get_level_values(0)
-                    units = optSum.loc[component].index.get_level_values(1)
-                    variables_unit = dict(zip(variables, units))
-
                     for variable in (
                         optSum.loc[component].index.get_level_values(0).unique()
                     ):
-                        df = optSum.loc[(component, variable)]
-                        if len(df.index.get_level_values(0).unique()) > 1:
-                            idx = df.index.get_level_values(0).unique()[-1]
-                            df = df.xs(idx, level=0)
+                        df_o = optSum.loc[(component, variable)]
+                        if "operation" in variable or variable=="operation":
+                            df = df_o.iloc[0:2,:].copy()
[…]
