### 2024-11-27 FZJ-IEK3-VSA/FINE#49 @GregorBecker
Hi @t-gross ,

the description of the examples looks sufficient. However, the units are still in square brackets and have indices.
This means that there is no reason for me to keep this issue open.
So feel free to close this issue and discuss if you want to make changes to the units.

Best regards.
@GregorBecker
