### 2024-11-27 FZJ-IEK3-VSA/FINE#50 @GregorBecker
Dear @t-gross,

the community guidelines look sufficient. Therefore, this issue can be closed.

Kind Regards,
@GregorBecker
