### 2024-11-27 FZJ-IEK3-VSA/FINE#51 @GregorBecker
Dear @t-gross,

as @leonardgoeke has already pointed out, it would be helpful to include the statement of need in the documentation. Otherwise, you have sufficiently addressed the comments on the paper, and the issue can be closed.

Kind regards,
@GregorBecker
