### 2024-11-27 FZJ-IEK3-VSA/FINE#51 @t-gross
Dear @GregorBecker,
thanks for your message.
I integrated the statement of need in the documentation (it will be found under "Purpose"). As soon as the merge request is done, I will close the issue.

Kind regards,
@t-gross
