### 2024-11-27 #6274 @t-gross
Thanks for the comments and latest feedback.
I implemented the statement of need in the documentation (it will be integrated under Purpose). This change and the first version of the mathematical description will be merged hopefully this week.
I will give a short update when the merge is done.
