### 2024-11-28 a91fd80
Integrate mathematical description and additional information into the documentation
