+++ docs/source/furtherReadingDoc.rst
@@ -13,8 +13,31 @@ Find examples for the application of FINE in the following published articles:
 * Caglayan et al. (2019): `The techno-economic potential of offshore wind energy with optimized future turbine designs in Europe <https://www.sciencedirect.com/science/article/abs/pii/S0306261919314813?via%3Dihub>`_ .
 
 
-* Kannengießer et al. (2020): `Reducing Computational Load for Mixed Integer Linear Programming: An Example for a District and an Island Energy System <https://www.mdpi.com/1996-1073/12/14/2825>`_ .
+* Kannengießer et al. (2020): `Reducing Computational Load for Mixed Integer Linear Programming: An Example for a District and an Island Energy System <https://doi.org/10.3390/en12142825>`_ .
 
 
 * Knosala et al. (2021): `Hybrid Hydrogen Home Storage for Decentralized Energy Autonomy <https://www.sciencedirect.com/science/article/pii/S0360319921013409>`_ .
 
+
+* Knosala et al. (2022): `The role of hydrogen in German residential buildings <https://doi.org/10.1016/j.enbuild.2022.112480>`_.
+
+
+* Spiller et al. (2022): `Reducing Carbon Emissions from the Tourist Accommodation Sector on Non-Interconnected Islands: A Case Study of a Medium-Sized Hotel in Rhodes, Greece <https://doi.org/10.3390/en15103801>`_.
+
+
+* Patil et al. (2022): `Advanced Spatial and Technological Aggregation Scheme for Energy System Models <https://doi.org/10.3390/en15249517>`_.
+
+
+* Singh et al. (2022): `Budget-cut: introduction to a budget based cutting-plane algorithm for capacity expansion models. <https://doi.org/10.1007/s11590-021-01826-w>`_.
+
+
+* Busch et al. (2023): `The role of liquid hydrogen in integrated energy systems–A case study for Germany <https://doi.org/10.1016/j.ijhydene.2023.05.308>`_.
+
+
+* Weinand et al. (2023): `Low-carbon lithium extraction makes deep geothermal plants cost-competitive in future energy systems <https://doi.org/10.1016/j.adapen.2023.100148>`_.
+
+
+* Franzmann et al. (2023): `Green hydrogen cost-potentials for global trade <https://doi.org/10.1016/j.ijhydene.2023.05.012>`_.
+
+
+* Jacob et al. (2023): `The future role of thermal energy storage in 100% renewable electricity systems <https://doi.org/10.1016/j.rset.2023.100059>`_:
