+++ docs/source/installationDoc.rst
@@ -1,4 +1,3 @@
-############
 Installation
 ############
 
