+++ docs/source/structureOfFINEDoc.rst
@@ -1,6 +1,5 @@
-**************************
 Python package description
-**************************
+##########################
 
 The python package description gives a general description of the code itself. It describes the package structure (modules and classes, with their functions and variables).
 
