### 2024-11-28 8337c3d
Integrate mathematical description and additional information into the documentation
