### 2024-11-29 #6274 @t-gross
Daer @arfon , dear reviewers,
The changes are implemented and merged to the develop branch. The documenation can be found here: https://vsa-fine.readthedocs.io/en/develop/
* The statement of need was integrated into the section Purpose
* The mathematical description was added unter User Guide>Mathematical Description

The changes should be merged into the master branch soon.
