### 2024-12-03 0dca6e0
Make capacityPerPlantUnit paramter dependent on investment period
