### 2024-12-03 ba090f5
Make capacityPerPlantUnit paramter dependent on investment period
