### 2024-12-09 #6274 @GregorBecker
Dear @arfon,

In my opinion, the latest changes fulfill the requirements for publication. I would just like to point out that the submitters have specified the branch **paper** and not **develop** as the branch to be published. I am not sure whether this will lead to a problem.

Kind regards
@GregorBecker
