### 2024-12-17 f3156d1
adds decommissioning based objective function contributions, use case implemented for materialSupplyPerCapacityDecommissioned, i.e. an trivial form of recycling, adds switch for second objective between material_demand and material_demand_and_supply
