+++ fine/component.py
@@ -49,6 +49,7 @@ class Component(metaclass=ABCMeta):
         stockCommissioning=None,
         floorTechnicalLifetime=True,
         materialDemandPerCapacity=None,
+        materialSupplyPerCapacityDecommissioned=None,
     ):
         """
         Constructor for creating an Component class instance.
@@ -687,6 +688,16 @@ class Component(metaclass=ABCMeta):
                 esM,
                 name,
                 materialDemandPerCapacity,
+                self.processedStockYears + esM.investmentPeriods,
+            )
+        )
+        # Recycling aka material "supply" based on decommissioning 
+        self.materialSupplyPerCapacityDecommissioned = materialSupplyPerCapacityDecommissioned
+        self.processed_materialSupplyPerCapacityDecommissioned = (
+            utils.checkAndSetMaterialDemandPerCapacity(
+                esM,
+                name,
+                materialSupplyPerCapacityDecommissioned,
                 self.processedStockYears + esM.investmentPeriods,
             )
         )
@@ -2892,9 +2903,7 @@ class ComponentModel(metaclass=ABCMeta):
         :param objective: type of objective for which the objective function contribution is returned. Default: costs
         :type objective: String
         """
-        # If new objective functions are implemented, they need to be added here!
-        if objective not in ["costs", "material_demand"]:
-            raise NotImplementedError("The chosen objective is not supported yet.")
+        
         if objective == "costs":
             capexCap = self.getEconomicsDesign(
                 pyM,
@@ -2932,9 +2941,12 @@ class ComponentModel(metaclass=ABCMeta):
             )
             objective_function_contribution = capexCap + capexDec + opexCap + opexDec
 
-        if objective == "material_demand":
+        if objective == "material_demand": 
+            # Only the MaterialDemandPerCapacity is considered
             objective_function_contribution = self.getMaterialDemand(esM, pyM)
-
+        if objective == "material_demand_and_supply":
+            # The second objective considers the MaterialDemandPerCapacity as well as the MaterialSupplyPerCapacityDecommissioned
+            objective_function_contribution = self.getMaterialDemand(esM, pyM) - self.getMaterialSupplyPerCapacityDecommissioned(esM, pyM)
         return objective_function_contribution
 
     def getMaterialDemand(self, esM, pyM, getOptValue=False):
@@ -2968,6 +2980,38 @@ class ComponentModel(metaclass=ABCMeta):
             return self.get_commissioning_based_objective_contribution(
                 esM, pyM, component_attribute
             )
+        
+    def getMaterialSupplyPerCapacityDecommissioned(self, esM, pyM, getOptValue=False):
+        """
+        Get material supply of a component when it is decommissioned. This can be used to build the
+        objective function or to get the optimal values after optimizing
+
+        :param esM: EnergySystemModel instance representing the energy system in which the component should be modeled.
+        :type esM: EnergySystemModel instance
+
+        :param pyM: pyomo ConcreteModel which stores the mathematical formulation of the model.
+        :type pyM: pyomo ConcreteModel
+
+        :param getOptValue: Boolean that defines the output of the function:
+
+            - True: Return the optimal material demand values.
+            - False: Return the material demand equation.
+
+            |br| * the default value is False.
+        :type getOptValue: boolean
+
+        :return: material demand equation or optimal material demand values (dictionary containing dataframes for each investment period)
+        :rtype: _type_
+        """
+        component_attribute = "processed_materialSupplyPerCapacityDecommissioned" 
+        if getOptValue:
+            return self.get_decommissioning_based_optimal_values(
+                esM, pyM, component_attribute
+            )
+        else:
[…]
