### 2025-01-07 afc7733
Allow timeseries for soc min/max and make dependent in investment period
