### 2025-01-07 2514d51
alternative approach to conversion of component dict to xarray for tranmisison.
