+++ test/IOManagement/test_xarrayio.py
@@ -120,7 +120,7 @@ def test_input_esm_to_netcdf_and_back(minimal_test_esM):
     """
 
     esm_original = deepcopy(minimal_test_esM)
-    xrIO.writeEnergySystemModelToNetCDF(esm_original, outputFilePath="test_esM.nc")
+    xrIO.writeEnergySystemModelToNetCDF(esm_original, outputFilePath="test_esM.nc", overwriteExisting=True)
     esm_from_netcdf = xrIO.readNetCDFtoEnergySystemModel(filePath="test_esM.nc")
 
     compare_esm_inputs(esm_original, esm_from_netcdf)
@@ -136,8 +136,7 @@ def test_output_esm_to_netcdf_and_back(minimal_test_esM):
 
     esm_original = deepcopy(minimal_test_esM)
     esm_original.optimize()
-
-    xrIO.writeEnergySystemModelToNetCDF(esm_original, outputFilePath="test_esM.nc")
+    xrIO.writeEnergySystemModelToNetCDF(esm_original, outputFilePath="test_esM.nc", overwriteExisting=True)
     esm_from_netcdf = xrIO.readNetCDFtoEnergySystemModel(filePath="test_esM.nc")
 
     compare_esm_inputs(esm_original, esm_from_netcdf)
