### 2025-01-07 829b101
added type checks for storage component parameters. also removed SOC...
