### 2025-01-08 666e370
Bug fix SOC Min/ Max Timeseries
