### 2025-01-09 6c477fa
Introduce endogenous technological learning
