### 2025-01-09 b6e46dc
remove ETL example
