### 2025-01-09 6c3d060
add error checks for simulataneous charge discharge
