+++ fine/conversion.py
@@ -53,6 +53,7 @@ class Conversion(Component):
         commissioningDependentCcf=False,
         emissionFactors=None,
         flowShares=None,
+        etlParameter=None
     ):
         # TODO: allow that the time series data or min/max/fixCapacity/eligibility is only specified for
         # TODO: eligible locations
@@ -258,6 +259,7 @@ class Conversion(Component):
             yearlyFullLoadHoursMin=yearlyFullLoadHoursMin,
             yearlyFullLoadHoursMax=yearlyFullLoadHoursMax,
             stockCommissioning=stockCommissioning,
+            etlParameter=etlParameter,
         )
 
         # opexPerOperation
