+++ fine/IOManagement/xarrayIO.py
@@ -45,9 +45,15 @@ def convertOptimizationInputToDatasets(esM, useProcessedValues=False):
             component: xr.Dataset() for component in component_dict[classname]
         }
 
+    # STEP 3.1 get _mapC for all transmission components
+    _mapC_dict = {}
+    for transmission_class in ["LinearOptimalPowerFlow", "Transmission"]:
+        for tech in component_dict[transmission_class].keys():
+            _mapC_dict[tech] = esM.getComponent(tech)._mapC
+            
     # STEP 4. Add all df variables to xr_ds
     xr_dss = utilsIO.addDFVariablesToXarray(
-        xr_dss, component_dict, df_iteration_dict, list(esM.locations)
+        xr_dss, component_dict, df_iteration_dict, _mapC_dict, list(esM.locations)
     )
 
     # STEP 5. Add all series variables to xr_ds
@@ -67,6 +73,26 @@ def convertOptimizationInputToDatasets(esM, useProcessedValues=False):
     attributes_xr.attrs = esm_dict
 
     xr_dss = {"Input": xr_dss, "Parameters": attributes_xr}
+
+    return xr_dss
+
+
+def convertPerformanceSummaryToDatasets(esM):
+    import pandas as pd
+
+    df = esM.performanceSummary.squeeze()
+    df = df.droplevel("Category")
+    df = df.apply(lambda x: pd.to_numeric(x, errors="ignore"))
+    # convert datetime to string
+    for idx, value in df.items():
+        if isinstance(value, pd.Timestamp):
+            print(value)
+            df.loc[idx] = value.strftime("%Y-%m-%d %H:%M:%S")
+    summary_dict = df.to_dict()
+    summary_xr = xr.Dataset()
+    summary_xr.attrs = summary_dict
+
+    xr_dss = {"PerformanceSummary": summary_xr}
 
     return xr_dss
 
@@ -382,7 +408,7 @@ def writeDatasetsToNetCDF(
             pass
 
     for group in datasets.keys():
-        if group == "Parameters":
+        if group == "Parameters" or group == "PerformanceSummary":
             xarray_dataset = datasets[group]
             _xarray_dataset = (
                 xarray_dataset.copy()
@@ -446,7 +472,6 @@ def writeDatasetsToNetCDF(
                 # Use mode='a' to append datasets to existing file. Variables will be overwritten.
                 mode=mode,
             )
-            continue
 
         elif group == "Results":
             for ip in datasets[group].keys():
@@ -468,7 +493,7 @@ def writeDatasetsToNetCDF(
                                 # Use zlib variable compression to reduce filesize with little performance loss
                                 # for our use-case. Complevel 9 for best compression.
                                 encoding={
-                                    var: {"zlib": True, "complevel": 9}
+                                    var: {"zlib": True, "complevel": 5}
                                     for var in list(
                                         datasets[group][ip][model][component].data_vars
                                     )
@@ -491,7 +516,7 @@ def writeDatasetsToNetCDF(
                             # Use zlib variable compression to reduce filesize with little performance loss
                             # for our use-case. Complevel 9 for best compression.
                             encoding={
-                                var: {"zlib": True, "complevel": 9}
+                                var: {"zlib": True, "complevel": 5}
                                 for var in list(
                                     datasets[group][model][component].data_vars
                                 )
@@ -989,6 +1014,12 @@ def writeEnergySystemModelToNetCDF(
     writeDatasetsToNetCDF(xr_dss_input, outputFilePath, groupPrefix=groupPrefix)
[…]
