+++ fine/utils.py
@@ -8,6 +8,15 @@ import fine as fn
 
 # ruff: noqa
 
+def isInRange(value, lowerBound, upperBound):
+    """Check if the input value is in the given range."""
+    if not (isinstance(value, float) or isinstance(value, int)):
+        raise TypeError("The input argument has to be a number")
+    if value <= upperBound and value >= lowerBound:
+        return value
+    else:
+        raise ValueError(f"The input argument has to be in the range [{lowerBound},{upperBound}]")
+
 def isString(string):
     """Check if the input argument is a string."""
     if not isinstance(string, str):
@@ -630,6 +639,19 @@ def checkInvestmentPeriodParameters(name, param, years):
             )
 
 
+def checkAndSetInvestmentPeriodParamters(name, param, esM):
+    checkInvestmentPeriodParameters(name, param, esM.investmentPeriodNames)
+    processedParam = {}
+    for ip in esM.investmentPeriods:
+        if param is None:
+            processedParam[ip] = None
+        if isinstance(param, dict):
+            processedParam[ip] = param[esM.investmentPeriodNames[ip]]
+        else:
+            processedParam[ip] = param
+    return processedParam
+
+
 def checkCapacityDevelopmentWithStock(
     investmentPeriods,
     capacityMax,
@@ -1035,9 +1057,14 @@ def checkAndSetInvestmentPeriodTimeSeries(
             parameter[ip] = checkAndSetTimeSeries(
                 esM, name, data[_ip], locationalEligibility, dimension
             )
+        elif isinstance(data, int) or isinstance(data, float):
+            _data = pd.DataFrame({loc: [data] * esM.numberOfTimeSteps for loc in esM.locations})
+            parameter[ip] = checkAndSetTimeSeries(
+                esM, name, _data, None, dimension
+            )
         else:
             raise TypeError(
-                f"Parameter of {name} should be a pandas dataframe or a dictionary."
+                f"Parameter of {name} does not match required type."
             )
     return parameter
 
@@ -1183,7 +1210,8 @@ def checkDesignVariableModelingParameters(
     if not isinstance(hasIsBuiltBinaryVariable, bool):
         raise TypeError("The hasCapacityVariable variable domain has to be a boolean.")
 
-    isStrictlyPositiveNumber(capacityPerPlantUnit)
+    for ip in esM.investmentPeriods:
+        isStrictlyPositiveNumber(capacityPerPlantUnit[ip])
 
     if not hasCapacityVariable and hasIsBuiltBinaryVariable:
         raise ValueError(
@@ -2862,4 +2890,94 @@ def checkAndSetFlowShares(comp, esM):
         }
 
     return processedFlowShares
+
+
+def getParametersForUnevenLifetimes(compName, loc, lifetimeAttr, esM):
+    ipEconomicLifetime = getattr(
+        esM.getComponent(compName), "ipEconomicLifetime"
+    )[loc]
+    ipTechnicalLifetime = getattr(
+        esM.getComponent(compName), "ipTechnicalLifetime"
+    )[loc]
+
+    # A) Fix operational costs for design variables.
+    # Fix operation costs are applied over the entire operational time.
+    # The duration of the operation time depends on the technical lifetime and
+    # (in case it is not a multiple of the interval) weather it is floored
+    # or ceiled to the next interval.
+    if lifetimeAttr == "ipTechnicalLifetime":
+        if esM.getComponent(compName).floorTechnicalLifetime:
+            intervalsWithCompleteCosts = math.floor(ipTechnicalLifetime)
+        else:
+            intervalsWithCompleteCosts = math.ceil(ipTechnicalLifetime)
[…]

[…]
