+++ paper/paper.md
@@ -32,54 +32,54 @@ affiliations:
     index: 1
   - name: RWTH Aachen University, Chair for Fuel Cells, Faculty of Mechanical Engineering, 52062 Aachen, Germany
     index: 2
-date: 09 January 2025
+date: 13 January 2025
 bibliography: FINE-literature.bib
 ---
 
 # Summary
 
-The decarbonization of energy systems worldwide requires a transformation of their design and operation across all sectors, that is, the residential and commercial, industrial, and transportation sectors. Energy system models are frequently employed for assessing these changes, providing scenarios on potential future system design and on how new technologies and a modified infrastructure will meet future energy demands. Thus, they support investment decisions and policy-making. The Python-based Framework for Integrated Energy System Assessment (`ETHOS.FINE`) is a software package that provides a toolbox for modeling, analyzing and evaluating such energy systems using mathematical optimization. 
+The decarbonization of energy systems worldwide requires a transformation in their design and operation across all sectors, including the residential and commercial, industrial, and transportation sectors. Energy system models are often used to assess these changes. These models provide scenarios for potential future system designs and show how new technologies and infrastructure will meet future energy demand. Thus, they support investment decisions and policy making. The Python-based Framework for Integrated Energy System Assessment (`ETHOS.FINE`) is a software package that provides a toolbox for modeling, analyzing and evaluating such energy systems using mathematical optimization. 
 
-`ETHOS.FINE` is not limited to a single instance of energy system. Instead, it can be freely adapted to consider multiple commodities, regions, time steps and investment periods. The optimization objective is to minimize the net present value of the system and is subject to technical and environmental constraints. If only one investment period is considered, the net present value equals the total annual costs of the system. The generic object-oriented implementation allows for arbitrary spatial scales and numbers of regions – from the local level, e.g., individual buildings, to the regional one, e.g., districts or industrial sites, to the national and international levels. 
+`ETHOS.FINE` is not limited to a single instance of energy systems. Instead, it can be freely adapted to consider multiple commodities, regions, time steps and investment periods. The optimization objective is to minimize the net present value of the system and is subject to technical and environmental constraints. If only one investment period is considered, the net present value equals the total annual costs of the system. The generic object-oriented implementation allows for arbitrary spatial scales and number of regions – from the local level, e.g., individual buildings, to the regional one, e.g., districts or industrial sites, to the national and international levels. 
 
-To reduce model size and complexity, the spatial technological resolution can be aggregated by built-in aggregation methods that are described in @Patil_2022. These methods include the aggregation of modeled regions and their given input parameters, and the grouping of technologies within the regions, e.g. to aggregate several time series for variable renewable energy sources. This also applies to the model's temporal resolution. Apart from using the full temporal resolution defined by the input data, integrated time series aggregation methods utilizing the built-in Python package tsam[^1] allow for reducing the model's complexity and its computation time [@Hoffmann_2022] while still enabling the flexibility of seasonal storage technologies, despite the reduced model complexity [@Kotzur_2018]. `ETHOS.FINE` supports the aggregation of time steps to typical periods, segmentation of the time series, and the combination of both. The aggregation methods, spatial and temporal aggregation, can directly be used in `ETHOS.FINE` by calling the corresponding functions. In addition, `ETHOS.FINE` enables the investigation of transformation pathways by considering multiple investment periods in a perfect foresight approach, as well as the stochastic optimization for a single year optimization with multiple sets of input parameters, e.g., changing energy demand forecasts or weather conditions, to find more robust energy system designs. 
[…]
