+++ paper/FINE-literature.bib
@@ -287,20 +287,21 @@ Type: https://doi.org/10.1016/j.rset.2023.100059},
 
 @phdthesis{Welder2022,
       author       = {Welder, Lara},
-      title        = {{O}ptimizing {C}ross-linked {I}nfrastructure for {F}uture
-                      {E}nergy {S}ystems},
+      othercontributors = {Stolten, Detlef and Görner, Klaus},
+      title        = {{O}ptimizing cross-linked infrastructure for future energy
+                      systems},
       volume       = {591},
       school       = {RWTH Aachen University},
       type         = {Dissertation},
       address      = {Jülich},
-      publisher    = {Forschungszentrum Jülich GmbH Zentralbibliothek Verlag},
-      reportid     = {FZJ-2022-06354},
-      isbn         = {978-3-95806-659-5},
-      series       = {Schriften des Forschungszentrums Jülich Reihe Energie $\&$
-                      Umwelt / Energy $\&$ Environment},
-      pages        = {xxiii, 360},
+      publisher    = {Forschungszentrum Jülich, Zentralbibliothek, Verlag},
+      reportid     = {RWTH-2022-11694},
+      series       = {Schriften des Forschungszentrums Jülich. Reihe Energie
+                      $\&$ Umwelt = Energy $\&$ environment},
       year         = {2022},
-      note         = {Dissertation, RWTH Aachen University, 2022},
+      note         = {Druckausgabe: 2022. - Onlineausgabe: 2022. - Auch
+                      veröffentlicht auf dem Publikationsserver der RWTH Aachen
+                      University 2023; Dissertation, RWTH Aachen University, 2022},
       abstract     = {Increasing shares of distributed, intermittent renewable
                       electricity generation as well as the need to decarbonize
                       all sectors in future energy systems require new
@@ -310,20 +311,56 @@ Type: https://doi.org/10.1016/j.rset.2023.100059},
                       electrolyzers will be deployed to create a renewable link
                       between the power sector and other sectors. Together with
                       technologies for power generation from renewably produced
-                      gases, electricity and gas infrastructurewill be strongly
+                      gases, electricity and gas infrastructure will be strongly
                       cross-linked. The design of such cross-linked infrastructure
                       is a computationally challenging undertaking and national
                       cross-linked infrastructure analyses are scarce in
-                      literature....},
-      cin          = {IEK-3},
-      cid          = {I:(DE-Juel1)IEK-3-20101013},
-      pnm          = {1111 - Effective System Transformation Pathways (POF4-111)
-                      / 1112 - Societally Feasible Transformation Pathways
-                      (POF4-111)},
-      pid          = {G:(DE-HGF)POF4-1111 / G:(DE-HGF)POF4-1112},
-      typ          = {PUB:(DE-HGF)3 / PUB:(DE-HGF)11},
-      urn          = {urn:nbn:de:0001-2023013134},
-      url          = {https://juser.fz-juelich.de/record/916578},
+                      literature. In this thesis, a new approach which enables the
+                      modeling of spatially and temporally resolved energy supply
+                      systems is proposed. In this context, the generic,
+                      open-source available modeling framework FINE is developed
+                      from scratch. FINE enables the optimal design and operation
+                      of cross-linked infrastructure levels. With its generic
+                      character, it can also be applied to model households,
+                      districts, one-nodal energy systems or international energy
+                      supply pathways and it has already been adapted by other
+                      scientists for these applications. By applying FINE,
+                      cross-linked infrastructure solutions for a German energy
+                      supply system in the year 2050 are investigated in this
[…]
