### 2025-01-16 09a2d21
ESM: add support for appsi_highs
if tee=True, the application deadlocks when used with multiprocessing
Set output logging to 0 to not spoil to much logging

Signed-off-by: Florian Maurer <maurer@fh-aachen.de>
