### 2025-01-16 #6274 @t-gross
Dear @arfon,
The archive is now uploaded and published on zenodo:
[https://doi.org/10.5281/zenodo.14673548](https://doi.org/10.5281/zenodo.14673548)

Paper has been checked for grammar and typos and updated in the branch [https://github.com/FZJ-IEK3-VSA/FINE/tree/paper](https://github.com/FZJ-IEK3-VSA/FINE/tree/paper).

Thank you.
