### 2025-01-20 #6274 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
